% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{accession}
\alias{accession}
\alias{SweepNames,PMRecording-method}
\title{Accession functions}
\usage{
\S4method{SweepNames}{PMRecording}(object)
}
\arguments{
\item{objectA}{\link[=PMRecording]{PMRecording} or PMCollection object}
}
\description{
These functions are used to access information from \link[=PMRecording]{PMRecording} and  \link[=PMCollection]{PMCollection} objects
}
