% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_atributes.R
\name{add_atributes}
\alias{add_atributes}
\title{add_atributes}
\usage{
add_atributes(tree, con, verbose = F, filename, con_pgf = con)
}
\description{
function inherited from ephys2 (/github/tdanker/ephys2/). Undocumented.
}
