% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddMetaData.R
\name{AddMetaData,PMRecording-method}
\alias{AddMetaData,PMRecording-method}
\title{Adds metadata to an PMCollection object}
\usage{
\S4method{AddMetaData}{PMRecording}(object, values, title = colnames(values))
}
\arguments{
\item{object}{A \link[=PMRecording]{PMRecording} or \link[=PMCollection]{PMCollection} object}

\item{values}{to be added}

\item{title}{title for metadata column(s).}
}
\value{
A \link[=PMRecording]{PMRecording} or \link[=PMCollection]{PMCollection}  object, respectively
}
\description{
This function adds metadata to an PMCollection object. The MetaData slot is a matrix with one row per sweep in PMRecording and one row per Series in PMCollections. The calling function is stored in the .MetaDataFx slot.
}
