% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddMetaData.R
\name{AddMetaData,PMSeries-method}
\alias{AddMetaData,PMSeries-method}
\title{Adds metadata to an PMExperiment object}
\usage{
\S4method{AddMetaData}{PMSeries}(object, values, title = colnames(values))
}
\arguments{
\item{object}{A \link[=PMSeries]{PMSeries} or \link[=PMExperiment]{PMExperiment} object}

\item{values}{to be added}

\item{title}{title for metadata column(s).}
}
\value{
A \link[=PMSeries]{PMSeries} or \link[=PMExperiment]{PMExperiment}  object, respectively
}
\description{
This function adds metadata to an PMExperiment object. The MetaData slot is a matrix with one row per sweep in PMSeries and one row per Series in PMExperiments. The calling function is stored in the .MetaDataFx slot.
}
