% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddMetaData.R
\name{AddMetaData,PRecording-method}
\alias{AddMetaData,PRecording-method}
\title{Adds metadata to an PCollection object}
\usage{
\S4method{AddMetaData}{PRecording}(object, values, title = colnames(values))
}
\arguments{
\item{object}{A \link[=PRecording]{PRecording} or \link[=PCollection]{PCollection} object}

\item{values}{to be added}

\item{title}{title for metadata column(s).}
}
\value{
A \link[=PRecording]{PRecording} or \link[=PCollection]{PCollection}  object, respectively
}
\description{
This function adds metadata to an PCollection object. The MetaData slot is a matrix with one row per sweep in PRecording and one row per Series in PCollections. The calling function is stored in the .MetaDataFx slot.
}
