% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddMetaData.R
\name{AddMetaData}
\alias{AddMetaData}
\title{Adds metadata to a \linkS4class{PRecording} or \linkS4class{PCollection} object}
\usage{
AddMetaData(object, values, title = colnames(values), Verbose = T)
}
\arguments{
\item{object}{A \linkS4class{PRecording} or \linkS4class{PCollection} object.}

\item{values}{The values to be added. Can be anything convertible into a \var{matrix}.}

\item{title}{The title(s) for the metadata column(s).}
}
\value{
A \linkS4class{PRecording} or \linkS4class{PCollection}   object, respectively
}
\description{
This function adds metadata to a PCollection object by filling the MetaData slot. The function calling \code{AddMetaData} is stored in the \var{.MetaDataFx} slot.
}
\details{
\strong{The MetaData slot} \cr
 In a \linkS4class{PRecording} object the MetaData slot is a \var{matrix} with each row corresponding to a sweep. \cr
 In a \linkS4class{PCollection} object the MetaData slot is a \var{matrix} with each row corresponding to a \linkS4class{PRecording} stored in the \linkS4class{PCollection}. \cr
 Column names must be unique.
}
\seealso{
\linkS4class{PRecording}, \linkS4class{PCollection}, \link[PatchR::apply]{apply()}, \link[PatchR::lapply]{lapply()}, \link[base::as.matrix]{as.matrix()}
}
