% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPMExperiment.R
\name{addPMExperiment,PMExperiment-method}
\alias{addPMExperiment,PMExperiment-method}
\title{Adds PMSeries to PMExperiment}
\usage{
\S4method{addPMExperiment}{PMExperiment}(
  object,
  PMSeries,
  Names = if (is.list(PMSeries)) {     lapply(PMSeries, function(x)
    x@RecordingParams@Filename) } else {     PMSeries@RecordingParams@Filename },
  Group = "Generic"
)
}
\arguments{
\item{object}{a PMExperiment object}

\item{PMSeries}{a PMSeries object, or list of PMSeries objects}

\item{Names}{names for imported PMSeries. Standard is variable name given in PMSeries if single PMSeries, i}

\item{Group}{Group variable or list thereof, of the same lenght as PMSeries}
}
\value{
A \link[=PMExperiment]{PMExperiment} object
}
\description{
Adds PMSeries to existing PMExperiment
}
