% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPMRecording.R
\name{addPMRecording,PMCollection-method}
\alias{addPMRecording,PMCollection-method}
\title{Adds PMRecording to PMCollection}
\usage{
\S4method{addPMRecording}{PMCollection}(
  object,
  PMRecording,
  Names = if (is.list(PMRecording)) {     lapply(PMRecording, function(x)
    x@RecordingParams@Filename) } else {     PMRecording@RecordingParams@Filename },
  Group = "Generic"
)
}
\arguments{
\item{object}{a PMCollection object}

\item{PMRecording}{a PMRecording object, or list of PMRecording objects}

\item{Names}{names for imported PMRecording. Standard is variable name given in PMRecording if single PMRecording, i}

\item{Group}{Group variable or list thereof, of the same lenght as PMRecording}
}
\value{
A \link[=PMCollection]{PMCollection} object
}
\description{
Adds PMRecording to existing PMCollection
}
