% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPMRecording.R
\name{addPMRecording}
\alias{addPMRecording}
\title{Adds PMRecording to PMCollection}
\usage{
addPMRecording(
  object,
  PMRecording,
  Names = if (is.list(PMRecording)) {     lapply(PMRecording, function(x)
    x@RecordingParams@Filename) } else {     PMRecording@RecordingParams@Filename },
  Group
)
}
\arguments{
\item{object}{a \linkS4class{PMRecording} object}

\item{PMRecording}{a PMRecording object, or list of PMRecording objects}

\item{Names}{names for imported PMRecording. Standard is variable name given in PMRecording if single PMRecording, i}

\item{Group}{Group variable or list thereof, of the same lenght as PMRecording}
}
\value{
A \linkS4class{PMCollection} object
}
\description{
Adds a \linkS4class{PMRecording} to existing \linkS4class{PMCollection}
}
