% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTrace.R
\docType{methods}
\name{addTrace,PMSeries-method}
\alias{addTrace,PMSeries-method}
\title{Adding data to PMSeries objects}
\usage{
\S4method{addTrace}{PMSeries}(object, Trace, Unit,
  Sweeps = colnames(mtx), isOrig = F, mtx)
}
\arguments{
\item{object}{A PMSeries object}

\item{Trace}{Name of the new Trace}

\item{Sweeps}{Names of the sweeps added. Default is colnames(mtx)}

\item{mtx}{Any object convertible into a matrix, that has the same dimension as data in the Data slot of object}
}
\description{
This function adds a new Trace (with data) from any object convertible into a matrix to the Data slot of PMSeries objects
}
