% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTrace.R
\name{AddTrace}
\alias{AddTrace}
\title{Add a trace/channel to PRecording objects}
\usage{
AddTrace(X, mtx, Trace, Unit, Sweeps = colnames(mtx), isOrig = F)
}
\arguments{
\item{X}{A \linkS4class{PRecording} object}

\item{mtx}{Any object convertible into a \var{matrix}. Rows are sampling points, columns are sweeps. Number of sampling points and sweeps must match the traces already present in \var{X}}

\item{Trace}{Name of the new trace}

\item{Unit}{The unit of the trace}

\item{Sweeps}{Names of the sweeps added. Must match the sweep names in \var{X}. Data in \var{Trace} will be sorted according to the order of sweeps in \var{X} before adding. Default is \code{colnames(X)}}

\item{isOrig}{if TRUE, marks added trace as an original recording.}
}
\value{
A matrix or \linkS4class{PRecording} X
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr
This function adds a new trace /channel from any object convertible into a matrix to an existing \linkS4class{PRecording} X.
}
\seealso{
\linkS4class{PRecording}, \linkS4class{PCollection}, \link[base:as.matrix]{as.matrix()}
}
