% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\name{apply,PMRecording-method}
\alias{apply,PMRecording-method}
\title{apply function for PMRecording objects}
\usage{
\S4method{apply}{PMRecording}(X, MARGIN, FUN, ReturnPMObject = F, Verbose = T)
}
\arguments{
\item{X}{a \linkS4class{PMRecording}  object}

\item{MARGIN}{a vector giving the subscripts which the function will be applied along. Understands "Time", "Sweep","Trace", or 1-3 resp.}

\item{FUN}{the function to be applied}

\item{ReturnPMObject}{whether to return results as a \linkS4class{PMRecording} with an additional, computed trace only works with \code{MARGIN=TRACE} or if applying \code{FUN} leaves dimensions unchanged. Default is \var{FALSE}, then returns a matrix.}
}
\value{
A \link[base::matrix]{matrix}  or \linkS4class{PMRecording}  object
}
\description{
This is the PatchMasteR analog to "apply"
}
