% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\docType{methods}
\name{apply,PMSeries-method}
\alias{apply,PMSeries-method}
\title{apply function for PMSeries objects}
\usage{
\S4method{apply}{PMSeries}(X, MARGIN, FUN, ReturnPMTRace = F)
}
\arguments{
\item{X}{a PMSeries object}

\item{MARGIN}{a vector giving the subscripts which the function will be applied along. Understands "Time", "Sweep","Trace", or 1-3 resp.}

\item{FUN}{the function to be applied}

\item{ReturnPMTRace}{whether to return results as a PMSeries with an additional, computed trace}
}
\value{
a matrix or PMSeries object
}
\description{
This is the PatchMastR analog to "apply"
}
