% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\name{apply,PRecording-method}
\alias{apply,PRecording-method}
\title{apply function for PRecording objects}
\usage{
\S4method{apply}{PRecording}(X, MARGIN, FUN, ReturnPMObject = F, Verbose = T)
}
\arguments{
\item{X}{A \linkS4class{PRecording}  object}

\item{MARGIN}{A vector giving the subscripts which the function will be applied along. Understands \code{Time}, \code{Sweep},\code{Trace}, or \code{1-3} resp.}

\item{FUN}{The function to be applied}

\item{ReturnPMObject}{Whether to return results as a \linkS4class{PRecording} with an additional, computed trace only works with \code{MARGIN="Trace"} or if applying \code{FUN} leaves dimensions unchanged. Default is \var{FALSE}, then returns a matrix.}

\item{Verbose}{Print what function is doing.}
}
\value{
A \link[base:matrix]{matrix}  or \linkS4class{PRecording} object.
}
\description{
This is the PatchR analog to \code{apply}
}
\seealso{
\link[base:apply]{base::apply()}, \link[PatchR:lapply]{PatchR::lapply()}
}
