% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.R
\name{as.data.frame,PMRecording-method}
\alias{as.data.frame,PMRecording-method}
\title{Converts PMRecording or PMCollection into a long data frame}
\usage{
\S4method{as.data.frame}{PMRecording}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{a \link[=PMRecording]{PMRecording} or \link[=PMCollection]{PMCollection}object}
}
\description{
Converts PMRecording or PMCollection into a data frame in a long representation, analog to tidyR's gather
}
