% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.R
\name{as.data.frame,PRecording-method}
\alias{as.data.frame,PRecording-method}
\title{Converts PRecording or PCollection into a long data frame}
\usage{
\S4method{as.data.frame}{PRecording}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{a \link[=PRecording]{PRecording} or \link[=PCollection]{PCollection}object}
}
\description{
Converts PRecording or PCollection into a data frame in a long representation, analog to tidyR's gather
}
