% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcode.R
\name{as_date}
\alias{as_date}
\title{Convert a character date to POSIXct}
\usage{
as_date(x, origin = "1899-12-30")
}
\arguments{
\item{x}{Character date}

\item{origin}{Origin date}
}
\value{
A POSIXct date
}
\description{
Convert a character date to POSIXct
}
