% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as_numeric}
\alias{as_numeric}
\title{Transcode a vector into numeric.}
\usage{
as_numeric(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A numeric vector.
}
\description{
Transcode a vector into numeric.
}
\keyword{internal}
