% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boltzmann.R
\name{Boltzmann,PRecording-method}
\alias{Boltzmann,PRecording-method}
\title{normalize true values to the predicted at min and max}
\usage{
\S4method{Boltzmann}{PRecording}(
  X,
  StimTrace = "V-mon",
  RespTrace = "I-mon",
  Time_Range_Stim = c(1.18, 1.19),
  Time_Range_Tail_Start = c(1.203, 1.204),
  Time_Range_Tail_End = c(1.33, 1.38),
  precision = 3,
  ReturnPObject = T,
  ...
)
}
\description{
normalize true values to the predicted at min and max
}
