% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boltzmann.R
\name{Boltzmann}
\alias{Boltzmann}
\alias{boltzmann_fit}
\title{Fit a Boltzmann function to data}
\usage{
Boltzmann(
  X,
  StimTrace = "V-mon",
  RespTrace = "I-mon",
  Time_Range_Stim = c(1.18, 1.19),
  Time_Range_Tail_Start = c(1.203, 1.204),
  Time_Range_Tail_End = c(1.33, 1.38),
  precision = 3,
  ReturnPObject = T,
  ...
)

boltzmann_fit(
  V,
  G,
  start_Vhlaf = V[which.min(abs(normalize(G) - 0.5))],
  start_k = estimate_slope(V, G),
  start_G_max = max(G) * 1.1,
  lower_Vhalf = min(V),
  lower_G_max = mean(G[seq(length(G) - round(length(G)/4), length(G))]) * 0.75,
  upper_Vhalf = max(V)
)
}
\arguments{
\item{X}{\linkS4class{PRecording} object}

\item{StimTrace}{The name of the \code{trace} that contains the stimulus}

\item{RespTrace}{The name of the \code{trace} that contains the response}

\item{Time_Range_Stim}{A numeric vector specifying the time range for the stimulus trace .}

\item{Time_Range_Tail_Start}{A numeric vector specifying the time range for measuring the tail current in the response trace.}

\item{Time_Range_Tail_End}{A numeric vector specifying the time range for measuring the leak current in the response trace (usually end of the voltage step eliciting the tail currents).}

\item{precision}{An integer specifying the precision for rounding the stimulus values.}

\item{ReturnPObject}{whether to return a \linkS4class{PRecording}/\linkS4class{PCollection} or a matrix.}

\item{V}{A numeric vector representing the voltage values.}

\item{G}{A numeric vector representing the conductance values.}

\item{start_k}{The initial estimate for k.
Defaults to the estimated slope of the data within a specified range using the \code{estimate_slope} function.}

\item{start_G_max}{The initial estimate for G_max.
Defaults to 1.1 times the maximum G value.}

\item{lower_Vhalf}{The lower bound for Vhalf.
Defaults to the minimum V value.}

\item{lower_G_max}{The lower bound for G_max.
Defaults to 0.75 times the mean of G values in the last quarter of the data.}

\item{upper_Vhalf}{The upper bound for Vhalf.
Defaults to the maximum V value.}

\item{start_Vhalf}{The initial estimate for Vhalf.
Defaults to the V value corresponding to G closest to 0.5 when normalized.}
}
\value{
For \code{Boltzmann()}: A \linkS4class{PRecording} or \linkS4class{PCollection}, respectively. In \linkS4class{PRecording} the \code{Fit} Slot will be added a list item termed \code{Boltzmann} containing the modeled fit and the \code{MetaData} Slot will contain additional columns as described below. In \linkS4class{PCollection} objects, a \code{V_half}  column will be added to the  \code{MetaData} Slot. If \code{ReturnPObject=F} a \link[base:data.frame]{data.frame} will be returned, as described below.
\describe{
\item{V}{The voltage values from the input data.}
\item{G}{The conductance values from the input data.}
\item{G.norm}{The normalized conductance values calculated using the fitted Boltzmann function.}
\item{G.predicted}{The predicted conductance values from the fitted Boltzmann function at the V values.}
\item{G.predicted.norm}{The normalized predicted conductance values from the fitted Boltzmann function at the V values.}
}
The \link[base:data.frame]{data.frame} also includes the following attributes:
\describe{
\item{Vhalf}{The estimated Vhalf parameter of the Boltzmann function.}
\item{k}{The estimated k parameter of the Boltzmann function.}
\item{G_max}{The estimated G_max parameter of the Boltzmann function.}
}

For \code{boltzmann_fit()}: An object of class "nls" representing the fitted Boltzmann function.
}
\description{
#' \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr
The \code{Boltzmann} method fits a Boltzmann function to data stored in a \linkS4class{PRecording} object. Therefore, values from the stimulus (usually voltage) and response (usually current) traces are averaged between the indicated time windows on a per-sweep basis. Internally, the \code{boltzmann_fit} function is called to perform the actual fitting.

The \code{boltzmann_fit} function fits a Boltzmann function to voltage and conductance data (V and G). It is normally called internally from the \code{Boltzmann} method to perform the actual fitting, but can also be called directly.
It estimates the parameters Vhalf, k, and G_max that best fit the data. The Boltzmann function is defined as: \deqn{(G_max - 0) / (1 + exp((Vhalf - V) / k))} and thus assumes that virtually all channels are in the closed state at the most negative voltage values.
}
\section{Functions}{
\itemize{
\item \code{boltzmann_fit}: Fit a Boltzmann function to voltage and conductance data.
}}

\examples{

######
                
V <- c(1, 2, 3, 4, 5)
G <- c(0.1, 0.3, 0.5, 0.7, 0.9)
fit <- boltzmann_fit(V, G)


}
