% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildCQPlot.R
\name{BuildQCPlot}
\alias{BuildQCPlot}
\title{Visualize quality metrics for PRecordings in PCollection}
\usage{
BuildQCPlot(object)
}
\arguments{
\item{object}{a \link[=PCollection]{PCollection} object object}
}
\value{
A \link[=PCollection]{PCollection} object
}
\description{
This function creates a basic visualization of quality-relevant recording parameters in the PCollection These will be stored in the Plots slot and can be accessed using \link[=Inspect]{Inspect}(object,"QC_Metrics")
}
