% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildSeriesPlot.R
\name{BuildSeriesPlot,PMRecording-method}
\alias{BuildSeriesPlot,PMRecording-method}
\title{Basic ggplots for  series data}
\usage{
\S4method{BuildSeriesPlot}{PMRecording}(object)
}
\arguments{
\item{object}{a \link[=PMRecording]{PMRecording} object object}
}
\value{
A \link[=PMRecording]{PMRecording} object
}
\description{
This function creats a basic visualization for each Trace in the object. These will be stored in the Plots slot
}
