% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildSeriesPlot.R
\name{BuildSeriesPlot,PMSeries-method}
\alias{BuildSeriesPlot,PMSeries-method}
\title{Basic ggplots for time series data}
\usage{
\S4method{BuildSeriesPlot}{PMSeries}(object)
}
\arguments{
\item{object}{a \link[=PMSeries]{PMSeries} object object}
}
\value{
A matrix or \link[=PMSeries]{PMSeries} object
}
\description{
This function creats a basic visualization for each Trace in the object. These will be stored in the Plots slot
}
