% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildSeriesPlot.R
\name{BuildSeriesPlot}
\alias{BuildSeriesPlot}
\title{Basic ggplots forseries data}
\usage{
BuildSeriesPlot(object)
}
\arguments{
\item{object}{a \link[=PMRecording]{PMRecording} object object}
}
\value{
A \link[=PMRecording]{PMRecording} object
}
\description{
This function creats a basic visualization for each Trace in the object. These will be stored in the Plots slot
}
