% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildTimeSeriesPlot.R
\docType{methods}
\name{BuildTimeSeriesPlot,PMTrace-method}
\alias{BuildTimeSeriesPlot,PMTrace-method}
\title{#' Basic ggplots for time series data}
\usage{
\S4method{BuildTimeSeriesPlot}{PMTrace}(object)
}
\arguments{
\item{object}{a PMTrace object}
}
\description{
This function creats a basic visualization for each channel in the object. These will be stored in the Plots slot
}
