% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_IV.R
\name{Calculate_IV,PMRecording-method}
\alias{Calculate_IV,PMRecording-method}
\title{Caluclates IV from a PMRecording or PMCollection object}
\usage{
\S4method{Calculate_IV}{PMRecording}(
  object,
  X_FROM,
  X_TO,
  ReturnPMObject = T,
  ITrace = "I-mon",
  VTrace = "V-mon",
  x_D_FROM = NA,
  X_D_TO = NA
)
}
\arguments{
\item{object}{a \link[=PMRecording]{PMRecording} object}

\item{X_FROM, X_TO, x_D_FROM, X_D_TO}{Time points to perform averaging for IV prodcution}

\item{ReturnPMObject}{whether to return results as a \link[=PMRecording]{PMRecording}  with an additional, computed trace. If set to \code{FALSE}, will return a \link[=base::matrix]{base::matrix}. Default is \code{TRUE}.}

\item{ITrace, VTrace}{Name of the traces containig Current(I) and and Voltage(V)}
}
\value{
a matrix or PMRecording with IV \link[=base::matrix]{base::matrix} and \link[=ggplot2::ggplot]{ggplot2::ggplot} stored in the MetaData and Plot slot, resp.
}
\description{
This function averages \link[=PMRecording]{PMRecording} objects by Trace, Sweep or Time. If object is a \link[=PMCollection]{PMCollection}. then does so for each Series stored in the object
}
