% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_IV.R
\name{Calculate_IV,PMSeries-method}
\alias{Calculate_IV,PMSeries-method}
\title{Caluclates IV from a PMSeries object}
\usage{
\S4method{Calculate_IV}{PMSeries}(
  object,
  X_FROM,
  X_TO,
  ReturnPMTRace = T,
  ITrace = "I-mon",
  VTrace = "V-mon",
  x_D_FROM = NA,
  X_D_TO = NA
)
}
\arguments{
\item{object}{a \link[=PMSeries]{PMSeries} object}

\item{X_FROM, X_TO, x_D_FROM, X_D_TO}{Time points to perform averaging for IV prodcution}

\item{ReturnPMTRace}{whether to return results as a \link[=PMSeries]{PMSeries}  with an additional, computed trace. If set to \code{FALSE}, will return a \link[=base::matrix]{base::matrix}. Default is \code{TRUE}.}

\item{ITrace, VTrace}{Name of the traces containig Current(I) and and Voltage(V)}
}
\value{
a matrix or PMSeries with IV \link[=base::matrix]{base::matrix} and \link[=ggplot2::ggplot]{ggplot2::ggplot} stored in the MetaData and Plot slot, resp.
}
\description{
This function averages \link[=PMSeries]{PMSeries} objects by Trace, Sweep or Time
}
