% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClearMetaData.R
\name{ClearMetaData}
\alias{ClearMetaData}
\title{Clear specified metadata columns from a PRecording or PCollection object}
\usage{
ClearMetaData(
  X,
  columns = try(colnames(GetMetaData(X)), silent = T),
  Verbose = T
)
}
\arguments{
\item{X}{A \linkS4class{PRecording} or \linkS4class{PCollection} object}

\item{columns}{A character vector specifying the columns to clear. Defaults to all columns in the MetaData slot.}

\item{Verbose}{Logical indicating whether to display warnings for non-existing columns. Defaults to \code{TRUE}.}
}
\value{
The modified \linkS4class{PRecording} or \linkS4class{PCollection} object with the specified columns cleared from the MetaData slot.
}
\description{
Clears the specified columns from the MetaData slot of a \code{PRecording} object.
}
\details{
This method clears the specified columns from the MetaData slot of a \linkS4class{PRecording} or \linkS4class{PCollection} object. If all columns are cleared,
the MetaData slot and the associated \code{.MetaDataFx} slot are reset to empty values. If only specific columns are cleared,
the corresponding columns in the MetaData slot are removed.
}
\section{Functions}{
\itemize{
\item \code{ClearMetaData}: This method clears the specified columns from the MetaData slot of a \code{PRecording} object.
}}

\examples{
# Create a PRecording object with MetaData
data("PRecording")
# add two columns of metadata
SampleData<-AddMetaData(SampleData,1:length(GetSweepNames(SampleData)),"ID1")
SampleData<-AddMetaData(SampleData,length(GetSweepNames(SampleData)):1,"ID2")
SampleData<-AddMetaData(SampleData,length(GetSweepNames(SampleData)):1,"ID3")
SampleData

# Clear the "ID1" and "ID2" columns from the MetaData slot
cleared_SampleData <- ClearMetaData(SampleData, columns = c("ID1", "ID2"))
cleared_SampleData

# Clear all columns from the MetaData slot
cleared_SampleData2 <- ClearMetaData(SampleData)
cleared_SampleData2

}
\seealso{
\linkS4class{PRecording} \linkS4class{PCollection}  \link{AddMetaData}
}
