% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenientScales.R
\name{convenientScales,PMRecording-method}
\alias{convenientScales,PMRecording-method}
\title{Makes PMRecording object scales convenient.}
\usage{
\S4method{convenientScales}{PMRecording}(object)
}
\arguments{
\item{object}{a \link[=PMRecording]{PMRecording} object}
}
\value{
a \link[=PMRecording]{PMRecording} object
}
\description{
This function converts scaling of a \link[=PMRecording]{PMRecording} or a \link[=PMCollection]{PMCollection} object and adjusts/adds SI prefixes
}
