% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenientScales.R
\name{convenientScales,PMSeries-method}
\alias{convenientScales,PMSeries-method}
\title{Makes PMSeries object scales convenient.}
\usage{
\S4method{convenientScales}{PMSeries}(object)
}
\arguments{
\item{object}{a \link[=PMSeries]{PMSeries} object}
}
\value{
a \link[=PMSeries]{PMSeries} object
}
\description{
This function converts scaling of a \link[=PMSeries]{PMSeries} object and adjusts/adds SI prefixes
}
