% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurrentDensity.R
\name{CurrentDensity,PMSeries-method}
\alias{CurrentDensity,PMSeries-method}
\title{Normalizes current traces to cell capacity}
\usage{
\S4method{CurrentDensity}{PMSeries}(object, trace = "I.mon")
}
\arguments{
\item{object}{a \link[=PMSeries]{PMSeries} object}
}
\value{
a \link[=PMSeries]{PMSeries} object
}
\description{
This function normalizes current traces to cell capacity and adds it as a new Trace
}
