% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{df_update}
\alias{df_update}
\title{Update a tibble values according to a second update tibble}
\usage{
df_update(data, update, by)
}
\arguments{
\item{data}{Original tibble}

\item{update}{Tibble used to patch the original tibble}

\item{by}{Columns used to join both tibbles

https://github.com/tidyverse/dplyr/issues/4595#issuecomment-548123242}
}
\description{
Update a tibble values according to a second update tibble
}
