% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatchR-data.R
\docType{data}
\name{.SampleData}
\alias{.SampleData}
\alias{SampleData}
\title{IV Protocol data}
\format{
An object of class \code{"PRecording"}; see \link[=PRecording]{PRecording}.
}
\usage{
data(PRecording)
}
\description{
This data set contains data from two channels (I-mon and V-mon) and three sweeps obtained using an IV recording protocol.
}
\examples{
data(PRecording)
SampleData<-BuildRecordingPlot(SampleData)
Inspect(SampleData, what = "I-mon")
}
\keyword{datasets}
