% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{downsample,PMRecording-method}
\alias{downsample,PMRecording-method}
\title{Donwsample a PMRecording object}
\usage{
\S4method{downsample}{PMRecording}(object, npnts.out = NULL, sample.rate = NULL)
}
\arguments{
\item{object}{a \link[=PMRecording]{PMRecording} object}

\item{npnts.out}{Number of points in output time axis, ignored if NULL}

\item{sample.rate}{New sampling rate in 1/[object@TimeUnit]}
}
\value{
A matrix or \link[=PMRecording]{PMRecording} object
}
\description{
Downsample \link[=PMRecording]{PMRecording} object along the time axis
}
