% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\docType{methods}
\name{downsample,PMTrace-method}
\alias{downsample,PMTrace-method}
\title{Donwsample a PMTrace object}
\usage{
\S4method{downsample}{PMTrace}(object, npnts.out = NULL,
  sample.rate = NULL)
}
\arguments{
\item{object}{a PMTrace}

\item{npnts.out}{Number of points in output matrix, ignored if NULL}

\item{sample.rate}{New sampling rate in 1/[object@TimeUnit]}
}
\description{
Downsample PMTrace object along the time axis
}
