% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{downsample,PRecording-method}
\alias{downsample,PRecording-method}
\title{Donwsample a PRecording object}
\usage{
\S4method{downsample}{PRecording}(object, npnts.out = NULL, sample.rate = NULL)
}
\arguments{
\item{object}{a \link[=PRecording]{PRecording} object}

\item{npnts.out}{Number of points in output time axis, ignored if NULL}

\item{sample.rate}{New sampling rate in 1/[object@TimeUnit]}
}
\value{
A matrix or \link[=PRecording]{PRecording} object
}
\description{
Downsample \link[=PRecording]{PRecording} object along the time axis
}
