% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropPMRecording.R
\name{dropPMRecording,PMCollection-method}
\alias{dropPMRecording,PMCollection-method}
\title{Drops PMRecording(s) from PMCollection}
\usage{
\S4method{dropPMRecording}{PMCollection}(object, PMRecording = NULL, Group = NULL)
}
\arguments{
\item{object}{a PMCollection object}

\item{PMRecording}{a name of a PMRecording object in the collection, or a list thereof}
}
\value{
A \link[=PMCollection]{PMCollection} object
}
\description{
Drops PMRecording(s) from PMCollection
}
