% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropPMRecording.R
\name{dropPRecording,PCollection-method}
\alias{dropPRecording,PCollection-method}
\title{Drops PRecording(s) from PCollection}
\usage{
\S4method{dropPRecording}{PCollection}(object, PRecording = NULL, Group = NULL)
}
\arguments{
\item{object}{a PCollection object}

\item{PRecording}{a name of a PRecording object in the collection, or a list thereof}
}
\value{
A \link[=PCollection]{PCollection} object
}
\description{
Drops PRecording(s) from PCollection
}
