% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropPMRecording.R
\name{DropPRecording}
\alias{DropPRecording}
\title{(OK) Drops Recording(s) from PCollection}
\usage{
DropPRecording(X, Recording = NULL, Group = NULL)
}
\arguments{
\item{X}{a PCollection X}

\item{Recording}{A name of a \linkS4class{PRecording} in the collection, or a list thereof}

\item{Group}{A name of A group in the\linkS4class{PCollection}. Drops all \linkS4class{PRecording} belonging to that group. Either Recording or Group need to be given}
}
\value{
A \linkS4class{PCollection} X
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr
Drops Recording(s) from PCollection
}
