% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boltzmann.R
\name{estimate_slope}
\alias{estimate_slope}
\title{Estimate the slope of a curve within specified margins}
\usage{
estimate_slope(V, G, margins = c(0.2, 0.8))
}
\arguments{
\item{V}{A numeric vector representing the independent variable.}

\item{G}{A numeric vector representing the dependent variable.}

\item{margins}{A numeric vector specifying the lower and upper margins for G values.
The default is c(0.2, 0.8).}
}
\value{
The estimated slope of the curve within the specified margins.
}
\description{
This function estimates the slope of a curve represented by two vectors, V and G,
within the specified margins. It calculates the difference in V values corresponding
to the range of G values within the margins, and divides it by the difference in
the margins to estimate the slope.
}
\keyword{internal}
