% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{filter_data_patch}
\alias{filter_data_patch}
\title{Filter a correspondance table according to additional colmuns within the table.}
\usage{
filter_data_patch(data, source = NULL, filtre = NULL)
}
\arguments{
\item{data}{A rename or recode data frame.}

\item{source}{A first filter with the source colname.}

\item{filtre}{A second filter with the filtre colname.}
}
\value{
A filtered data frame.
}
\description{
Filter a correspondance table according to additional colmuns within the table.
}
