% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcode.R
\name{fix_birth_date}
\alias{fix_birth_date}
\title{Fix birth date}
\usage{
fix_birth_date(date)
}
\arguments{
\item{date}{Date to fix}
}
\value{
A birth date
}
\description{
Treats a date as a birth date for years equal or prior to 1969.
}
