% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get}
\alias{get}
\alias{getSweepNames,PMRecording-method}
\alias{getSweepNames,PMCollection-method}
\alias{getTraceNames,PMRecording-method}
\alias{getTimeTrace,PMRecording-method}
\alias{getTimeTrace,PMCollection-method}
\alias{getCSlow,PMRecording-method}
\alias{getCSlow,PMCollection-method}
\alias{getGroupMembers,PMCollection-method}
\alias{getRecParam,PMRecording-method}
\alias{getRecParam,PMCollection-method}
\alias{getMetaData,PMRecording-method}
\alias{getMetaData,PMCollection-method}
\title{Accession functions}
\usage{
\S4method{getSweepNames}{PMRecording}(object)

\S4method{getSweepNames}{PMCollection}(object)

\S4method{getTraceNames}{PMRecording}(object)

\S4method{getTimeTrace}{PMRecording}(object)

\S4method{getTimeTrace}{PMCollection}(object)

\S4method{getCSlow}{PMRecording}(object)

\S4method{getCSlow}{PMCollection}(object)

\S4method{getGroupMembers}{PMCollection}(object, which)

\S4method{getRecParam}{PMRecording}(object, which)

\S4method{getRecParam}{PMCollection}(object, which)

\S4method{getMetaData}{PMRecording}(object, which = colnames(object@MetaData))

\S4method{getMetaData}{PMCollection}(object, which = colnames(object@MetaData))
}
\arguments{
\item{object}{A \linkS4class{PMRecording} or \linkS4class{PMCollection}  object}

\item{which}{A name of a valid Group in a \linkS4class{PMCollection} (for \code{getGroupMembers}). \cr A name or a vector of names of slot(s) in RecordingParams (for \code{getRecParam}). \cr A name or a vector of names of a column in the MetaData slot (for \code{getMetaData}).}
}
\description{
These functions are used to access information from \linkS4class{PMRecording} and/or \linkS4class{PMCollection} objects
}
