% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\docType{methods}
\name{getChannels,PMTrace-method}
\alias{getChannels,PMTrace-method}
\title{getChannels}
\usage{
\S4method{getChannels}{PMTrace}(object)
}
\arguments{
\item{object}{A PMTrace object}
}
\description{
get list of Channels from a PMTrace object
}
