% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\docType{methods}
\name{getCs,PMSeries-method}
\alias{getCs,PMSeries-method}
\title{getCs}
\usage{
\S4method{getCs}{PMSeries}(object)
}
\arguments{
\item{object}{A PMSeries object}
}
\description{
get Cs from PMSeries
}
