% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{GetData,PCollection-method}
\alias{GetData,PCollection-method}
\title{FIXME:
tmp<-GetData(DN81_MoL,Group = "Kv7.2_Kv8.1")}
\usage{
\S4method{GetData}{PCollection}(
  X,
  Traces = GetTraceNames(X),
  Sweeps = GetSweepNames(X),
  Time = range(GetTimeTrace(X)),
  Recordings = GetRecordingNames(X),
  Group = GetGroupNames(X),
  TimeExclusive = F,
  nowarnings = F
)
}
\description{
Error in validObject(.Object) :

invalid class “PCollection” object: invalid object for slot "MetaData" in
class "PCollection": got class "logical", should be or extend class "matrix"
}
