% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getGroupMembers,PMCollection-method}
\alias{getGroupMembers,PMCollection-method}
\title{getGroupMembers}
\usage{
\S4method{getGroupMembers}{PMCollection}(object, Group)
}
\arguments{
\item{object}{A PMCollection object}

\item{Goup}{one or more name(s) of Groups contained in PMCollection}
}
\description{
get names of all memebers of a group from PMCollection
}
