% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getMetaData,PMRecording-method}
\alias{getMetaData,PMRecording-method}
\title{getMetaData}
\usage{
\S4method{getMetaData}{PMRecording}(object, which = colnames(object@MetaData))
}
\arguments{
\item{object}{A PMRecording or PMCollection object}

\item{which}{columns form MetaData to retrieve. Default is all.}
}
\description{
get MetaData from PMRecording or PMCollection
}
