% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeries.R
\name{getSeries}
\alias{getSeries}
\title{getSeries}
\usage{
getSeries(node_or_tree, file = NA, exp = NA, ser = NA, trace = 1,
  read_data = T, CompoundName = NA)
}
\description{
function inherited from ephys2 (/github/tdanker/ephys2/). Undocumented.
}
