% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeries_from_node.R
\name{getSeries_from_node}
\alias{getSeries_from_node}
\title{getSeries_from_node}
\usage{
getSeries_from_node(node, sweeps = 1:length(node), trace = 1,
  read_data = T)
}
\description{
function inherited from ephys2 (/github/tdanker/ephys2/). Undocumented.
}
