% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStimName_from_unique_seriesName.R
\name{getStimName_from_unique_seriesName}
\alias{getStimName_from_unique_seriesName}
\title{getStimName_from_unique_seriesName}
\usage{
getStimName_from_unique_seriesName(string)
}
\description{
function inherited from ephys2 (/github/tdanker/ephys2/). Undocumented.
}
