% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\docType{methods}
\name{getSweeps,PMSeries-method}
\alias{getSweeps,PMSeries-method}
\title{getSweeps}
\usage{
\S4method{getSweeps}{PMSeries}(object)
}
\arguments{
\item{object}{A PMSeries object}
}
\description{
get list of Sweeps from a PMSeries object
}
