% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\docType{methods}
\name{getSweeps,PMTrace-method}
\alias{getSweeps,PMTrace-method}
\title{getSweeps}
\usage{
\S4method{getSweeps}{PMTrace}(object)
}
\arguments{
\item{object}{A PMTrace object}
}
\description{
get list of Sweeps from a PMTrace object
}
