% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getTimeTrace,PMRecording-method}
\alias{getTimeTrace,PMRecording-method}
\title{getTimeTrace}
\usage{
\S4method{getTimeTrace}{PMRecording}(object)
}
\arguments{
\item{object}{A \link[=PMRecording]{PMRecording}  object}
}
\description{
get Time trace from a \link[=PMRecording]{PMRecording}  object
}
