% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\docType{methods}
\name{getTimeTrace,PMSeries-method}
\alias{getTimeTrace,PMSeries-method}
\title{getTimeTrace}
\usage{
\S4method{getTimeTrace}{PMSeries}(object)
}
\arguments{
\item{object}{A PMSeries object}
}
\description{
get Time trace from a PMSeries object
}
