% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\docType{methods}
\name{getTimeTrace,PMTrace-method}
\alias{getTimeTrace,PMTrace-method}
\title{getTimeTrace}
\usage{
\S4method{getTimeTrace}{PMTrace}(object)
}
\arguments{
\item{object}{A PMTrace object}
}
\description{
get Time trace from a PMTrace object
}
