% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrace_.R
\name{getTrace_}
\alias{getTrace_}
\title{getTrace_}
\usage{
getTrace_(con, ptr, start = 0, n = NA, read_data = T, name = "",
  con_dat = con)
}
\description{
function inherited from ephys2 (/github/tdanker/ephys2/). Undocumented.
}
