% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getTraces,PMSeries-method}
\alias{getTraces,PMSeries-method}
\title{getTraces}
\usage{
\S4method{getTraces}{PMSeries}(object)
}
\arguments{
\item{object}{A \link[=PMSeries]{PMSeries} or \link[=PMRecordingParams]{PMRecordingParams} object}
}
\description{
get list of Traces from a \link[=PMSeries]{PMSeries} or \link[=PMRecordingParams]{PMRecordingParams} object
}
