% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\docType{methods}
\name{getTraces,PMTrace-method}
\alias{getTraces,PMTrace-method}
\title{getTraces}
\usage{
\S4method{getTraces}{PMTrace}(object)
}
\arguments{
\item{object}{A PMTrace or PMRecordingParams object}
}
\description{
get list of Traces from a PMTrace object
}
