% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportCollection.R
\name{ImportCollection}
\alias{ImportCollection}
\title{Imports a collection of electrophysiology recordings}
\usage{
ImportCollection(
  filelist = NULL,
  traces = c(1, 2),
  parent.dir = "",
  filetype = NULL,
  verbose = FALSE,
  encoding = getOption("encoding")
)
}
\arguments{
\item{filelist}{A data.frame. If importing from \href{https://www.heka.com/downloads/downloads_main.html#down_patchmaster}{Patch Master .dat} files, it must contain the following three columns: "FileName","Experiment","Series". FileName can be an absolute path or relative to \var{parent.dir}. "Experiment","Series" correspond to the second and third order hierarchy in the PatchMaster file tree. If importing from \href{https://swharden.com/pyabf/abf2-file-format/}{Axon Binary (.abf)} files, a single column "FileName" is sufficient. An optional column is "Group" and can be used to assign the individual recordings into groups}

\item{traces}{Index or indices of Traces/channels to import. The trace is the 3rd order hierarchy in PatchMaster file tree. Must be vector of numeric of length > 0. Default is \code{c(1,2)}.}

\item{parent.dir}{Parent directory where to search for FileNames from \code{filelist}. Optional if no full path is provided in \var{filelist}.}

\item{filetype}{The type of the files to be imported. Accepted values are 'PatchMaster' or 'ABF' If left blank (default), it will be guessed from the extension of the first file in list}

\item{verbose}{Whether to display verbose messages during the import process. Default is \code{FALSE}.}

\item{encoding}{File encoding to use, default is \code{getOption("encoding")}}
}
\value{
A \linkS4class{PCollection} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr
Imports a collection of electrophysiology recordings. All files in one condition must be acquired using the same recording protocol. I.e. the number of traces/channels, data points per sweep and number of sweeps must be identical.
}
\examples{
\dontrun{
rec.df<-data.frame("/path/to/file.dat",1,1)
colnames(rec.df)<-c(("FileName","Experiment","Series"))
tmp<-ImportCollection(rec.df,traces = c(1,2))
}
}
\seealso{
\linkS4class{PRecording}, \link[=ImportPRecording]{ImportPRecording}
}
