% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportPMSeries.R
\docType{methods}
\name{ImportPMSeries,PMTrace-method}
\alias{ImportPMSeries,PMTrace-method}
\title{Import Series from Patch Master *.dat file into a PMTrace object}
\usage{
\S4method{ImportPMSeries}{PMTrace}(filename, experiment = 1,
  series = 1, traces = c(1, 2))
}
\arguments{
\item{filename}{path to a Patch Master *.dat file.}

\item{experiment}{Index of Experiment to import (1st order hirarchy in PM dat file).}

\item{series}{Index of Series to import (2nd order hirarchy in PM dat file).}

\item{traces}{Traces to import (indices). Must be vector of numerics of length > 0.}
}
\description{
This function imports Series from Patch Master *.dat files and creates PMTrace objects
}
