% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectSeries.R
\name{Inspect,PMRecording-method}
\alias{Inspect,PMRecording-method}
\title{Inspect}
\usage{
\S4method{Inspect}{PMRecording}(object, what = "I.mon")
}
\arguments{
\item{object}{a \link[=base::matrix]{base::matrix} object}

\item{what}{Specifies a plot to be plotted. any ggplot inside the Plots slot can be named}
}
\description{
This function plots an interactive \link[=plotly]{plotly} graph from a plot generated using a Build method.
}
