% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inspect.R
\name{Inspect}
\alias{Inspect}
\title{Interactively view Plots}
\usage{
Inspect(X, what = "I.mon")
}
\arguments{
\item{X}{A \linkS4class{PRecording} or \linkS4class{PCollection} object}

\item{what}{Specifies a plot to be plotted. any ggplot inside the Plots slot can be named}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr
This function plots an interactive \link[=plotly]{plotly} graph from a plot sored in a \linkS4class{PCollection} or \linkS4class{PRecording} object generated using a Build* method.
}
\examples{
# Build a stimulus response curve from a single recording. Here, a current
# response, averaged from 0.8-1s of the trace is used. A PRecording object with
# an updated slot Plots is returned.
data(PRecording)
SampleData<-BuildStimRespPlot(
  SampleData,
  StimTrace = "V-mon",
  RespTrace = "I-mon",
  Time = c(0.8,1),
  fun = mean,
  ReturnPObject = TRUE
)
# Interactively analyse graph
Inspect(SampleData,"StimRespPlot")
}
\seealso{
\link[=Plot]{Plot}, \link[=Plot]{GetPlot}, \link[=Plot]{GetPlotNames}
}
