% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectTimeSeries.R
\name{InspectSeries}
\alias{InspectSeries}
\title{InspectSeries}
\usage{
\S4method{InspectTimeSeries}{PMSeries}(object, Trace = "I.mon")
}
\arguments{
\item{object}{a \link[=base::matrix]{base::matrix} object}

\item{Trace}{Specifies a trace to be plotted}
}
\description{
This function plots an interactive \link[=plotly]{plotly} graph from a plot generated using \link[=BuildSeriesPlot.]{BuildSeriesPlot}
}
