% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectTimeSeries.R
\docType{methods}
\name{inspectTimeSeries}
\alias{inspectTimeSeries}
\title{inspectTimeSeries}
\usage{
\S4method{inspectTimeSeries}{PMTrace}(object, Channel = "I.mon")
}
\arguments{
\item{object}{a PMTrace object}

\item{Channel}{channel to plot}
}
\description{
This function creates an interactive graph from a plot generated using BuildTimeSeriesPlot
}
