% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapply.R
\name{lapply,PCollection-method}
\alias{lapply,PCollection-method}
\title{apply function for PCollection objects}
\usage{
\S4method{lapply}{PCollection}(X, FUN, ReturnPObject = F)
}
\arguments{
\item{X}{a \linkS4class{PCollection} object}

\item{FUN}{the function to be applied}

\item{ReturnPObject}{whether to return results as a \linkS4class{PCollection}  Default is \var{FALSE}, then returns a list}
}
\value{
A \linkS4class{PCollection} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr
This is the PatchMastR analog to "lapply" for PCollection objects
}
