% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapply.R
\name{lapply,PMExperiment-method}
\alias{lapply,PMExperiment-method}
\title{apply function for PMExperiment objects}
\usage{
\S4method{lapply}{PMExperiment}(X, FUN, ReturnPMExperiment = F)
}
\arguments{
\item{X}{a PMExperiment object}

\item{FUN}{the function to be applied}

\item{ReturnPMExperiment}{whether to return results as a PMExperiment. Default is \var{FALSE}, then returns a list}
}
\value{
A \link[=PMExperiment]{PMExperiment} object
}
\description{
This is the PatchMastR analog to "lapply" for PMExperiment objects
}
