% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{length}
\alias{length}
\title{Length of an object}
\arguments{
\item{x}{a \var{PRecording} or \var{PCollection} object.}
}
\value{
The number of recordings in a \linkS4class{PCollection} or the number of sweeps from a \linkS4class{PRecording} as a one-dimensional \var{numeric} vector.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr
Returns the number of recordings in a \linkS4class{PCollection} or the number of sweeps from a \linkS4class{PRecording}
}
