% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_maj.R
\name{liste_champs_unique}
\alias{liste_champs_unique}
\title{Retourne la liste complete et dedoublonnee des champs de data frame}
\usage{
liste_champs_unique(import, table_rename, fichier_csv = "champs_unique.csv")
}
\arguments{
\item{import}{Une liste de data frame.}

\item{table_rename}{Une table de correspondance entre anciens et nouveaux noms de champ.}

\item{fichier_csv}{Un chemin de fichier csv en sortie.}
}
\value{
Un vecteur de noms de champ.
}
\description{
Retourne la liste complète et dédoublonnée des champs de data frame
}
