% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_maj.R
\name{maj_source}
\alias{maj_source}
\title{Met a jour/Harmonise une liste de data frame et les compile dans un data frame unique}
\usage{
maj_source(import, champs_filtre = NULL, ligne_fin = NULL)
}
\arguments{
\item{import}{Une liste de data frame.}

\item{champs_filtre}{Un nom de champ pour lequel les valeurs vides sont supprimées.}

\item{filtre_expression}{Une expression de type character supprimant des lignes des data frame.}
}
\value{
Un data frame unique
}
\description{
Met à jour/Harmonise une liste de data frame et les compile dans un data frame unique.
}

