% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SweepStats.R
\name{MeasureSweeps,PMCollection-method}
\alias{MeasureSweeps,PMCollection-method}
\title{Perform measurements on all recordings in the collection on a per-sweep basis.}
\usage{
\S4method{MeasureSweeps}{PMCollection}(
  object,
  Trace = "I-mon",
  Sweeps = SweepNames(object),
  Time,
  label,
  fun = mean
)
}
\arguments{
\item{object}{a \link[=PMCollection]{PMCollection} object}

\item{Trace}{Trace to perform Stats on}

\item{Sweeps}{List of Sweeps to process}

\item{Time}{either a range of time points to keep, or, if \code{fun} is a binary operator then two particular time points}

\item{label}{A label (if \code{Sweeps} has length 1) or a prefix}

\item{fun}{function to apply on sweep. Can be anything that woks with \link[=base::apply]{apply}. But will be usuall \link[=base::mean]{mean}, \link[=base::max]{max}, \link[=base::min]{min}, or \link[=base::`-`]{`-`}.}
}
\value{
A \link[=PMRecording]{PMRecording} object with an updated MetaData Slot.
}
\description{
This function performs measurements on all recordings in the collection for a given trace on a per-sweep basis. This can be useful e.g. for generating data underlying a Time Series Plot.
}
