% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newPCollection.R
\name{NewPCollection}
\alias{NewPCollection}
\title{(OK) Generates a new PCollection from a single PRecording}
\usage{
NewPCollection(PRecording, Names = NULL, Group = "Generic")
}
\arguments{
\item{PRecording}{a PRecording object, or list of PRecording objects}

\item{Names}{Names of the PRecording, character vector w same number of entires as PRecording}

\item{Group}{Group variable or list thereof, of the same lenght as PRecording}
}
\value{
A \linkS4class{PCollection}object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr
Generates a new \linkS4class{PCollection} from a single \linkS4class{PRecording}
}
