% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newPMexperiment.R
\name{newPMCollection}
\alias{newPMCollection}
\title{Generates a new PMCollection from a single PMRecording}
\usage{
newPMCollection(PMRecording, Names = NULL, Group = "Generic")
}
\arguments{
\item{PMRecording}{a PMRecording object, or list of PMRecording objects}

\item{Names}{Names of the PMRecording, character vector w same number of entires as PMRecording}

\item{Group}{Group variable or list thereof, of the same lenght as PMRecording}
}
\value{
A \link[=PMCollection]{PMCollection} object
}
\description{
Generates a new PMCollection from a single PMRecording
}
