% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newPMexperiment.R
\name{newPMExperiment}
\alias{newPMExperiment}
\title{Generates a new PMExperiment from a single PMSeries}
\usage{
newPMExperiment(PMSeries, Names = NULL, Group = "Generic")
}
\arguments{
\item{PMSeries}{a PMSeries object, or list of PMSeries objects}

\item{Group}{Group variable or list thereof, of the same lenght as PMSeries}
}
\value{
A \link[=PMExperiment]{PMExperiment} object
}
\description{
Generates a new PMExperiment from a single PMSeries
}
