% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{normalise_colnames}
\alias{normalise_colnames}
\title{Normalise column names of a data frame.}
\usage{
normalise_colnames(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A data frame with normalised colmun names.\cr
}
\description{
Normalise column names of a data frame.
}
\examples{
data <- dplyr::data_frame(
  "Type d'unité Sirus : entreprise profilée ou unité légale" = NA_character_,
  "Nic du siège"= NA_character_
)
patchr::normalise_colnames(data)

}
