% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize a numeric vector to a specified range}
\usage{
normalize(x, range = c(0, 1))
}
\arguments{
\item{x}{A numeric vector to be normalized.}

\item{range}{A numeric vector specifying the desired range. The default range is \link{0, 1}.}
}
\value{
A numeric vector with values normalized to the specified range.
}
\description{
The \code{normalize} function takes a numeric vector and normalizes its values
to a specified range.
}
\examples{
x <- c(1, 2, 3, 4, 5)
normalize(x)  # Normalize x to the default range [0, 1]
normalize(x, range = c(0, 100))  # Normalize x to the range [0, 100]

}
\seealso{
\code{\link{scale}}
}
