% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalize.R
\name{Normalize_PCollection}
\alias{Normalize_PCollection}
\title{FIXME: TURN INTO METHOD FOR both PObjects and document
turn into scale function allow for scale "toMax", "MintoMax"}
\usage{
Normalize_PCollection(object, Trace)
}
\description{
FIXME: TURN INTO METHOD FOR both PObjects and document
turn into scale function allow for scale "toMax", "MintoMax"
}
