% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatchR-package.R
\docType{package}
\name{PatchR-package}
\alias{PatchR}
\alias{PatchR-package}
\title{PatchR: An environment for working with ephys (PatchClamp) recordings}
\description{
An environment for working with ephys (PatchClamp) recordings. Currently import for HE(OK)A's PatchMaster is implemented.
}
\keyword{internal}
