% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PCollection.R
\docType{class}
\name{PCollection-class}
\alias{PCollection-class}
\alias{PCollection}
\title{S4 class storing a collection of ePhys Treaces}
\description{
This class stores a collection of \linkS4class{PRecording}s in a single object. Facilitates identical processing of related of recordings.
}
\details{
\describe{
   \item{Series}{A list containing individual PRecording objects of identical dimensisons, ProtocolName, RecMode and trace names}
   \item{Names}{Name identifier for the individual Series. Unique.}
   \item{Group}{Group identifier for the individual Series.}
   \item{MetaData}{A matrix with metatdata, e.g. calculated summay stats of each PRecording object. One row per entry in Series allowed}
   \item{Plots}{A slot to store ggplots}
   \item{RecordingParams}{Stores the Recording parameters that must be identical for all entries in Series (ProtocolName, RecMode, Traces)}
 }
}
\seealso{
\linkS4class{PRecording}
}
