% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1Plot.R, R/BuildQCPlot.R,
%   R/BuildRecordingPlot.R
\name{Plot}
\alias{Plot}
\alias{BuildStimRespPlot}
\alias{BuildTimeSeriesPlot}
\alias{BuildGroupComparisonPlot}
\alias{BuildQCPlot}
\alias{BuildRecordingPlot}
\title{Plotting methods}
\usage{
BuildStimRespPlot(
  X,
  StimTrace = "V-mon",
  RespTrace = "I-mon",
  Time,
  fun = mean,
  ReturnPObject = T
)

BuildTimeSeriesPlot(
  X,
  RespTrace = "I-mon",
  Time,
  fun = mean,
  ReturnPObject = T
)

BuildGroupComparisonPlot(
  X,
  Sweep,
  RespTrace = "I-mon",
  Time,
  fun = mean,
  ReturnPObject = T
)

BuildQCPlot(X, ReturnPObject = T)

BuildRecordingPlot(X)
}
\arguments{
\item{X}{A \linkS4class{PRecording} or \linkS4class{PCollection} object}

\item{StimTrace}{The name of the \code{trace} that contains the stimulus}

\item{RespTrace}{The name of the \code{trace} that contains the response}

\item{Time}{either a range of time points to keep, or, if \code{TimeExclusive} is \code{TRUE}, then two particular time points}

\item{fun}{Function to apply on graph for stimulus response plotting}

\item{ReturnPObject}{whether to return a \linkS4class{PRecording}/\linkS4class{PCollection} or a matrix.}

\item{Sweep}{Sweep to analyse for group comparison}
}
\value{
A \linkS4class{PCollection} with an item added to the Plots slot if \code{ReturnPObject=T} or a \link[ggplot2:ggplot]{ggplot}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr
These methods create "typical" electrophysiology graphs, like stimulus-response curves, time-series or point statistics.
}
\section{Functions}{
\itemize{
\item \code{BuildStimRespPlot}: This method builds a stimulus-response curve

\item \code{BuildTimeSeriesPlot}: This method builds a time Series plot

\item \code{BuildGroupComparisonPlot}: This method builds a boxplot for comparison between groups as stored in the \linkS4class{PCollection}.

\item \code{BuildQCPlot}: This function creates a basic visualization of quality-relevant recording parameters in the PCollection

\item \code{BuildRecordingPlot}: This function creates a basic visualization for each Trace in a single \linkS4class{PRecording} object.
}}

\examples{
# Build a stimulus response curve from a single recording. Here, a current
# response, averaged from 0.8-1s of the trace is used. A PRecording object with
# an updated slot Plots is returned.
data(PRecording)
SampleData<-BuildStimRespPlot(
  SampleData,
  StimTrace = "V-mon",
  RespTrace = "I-mon",
  Time = c(0.8,1),
  fun = mean,
  ReturnPObject = TRUE
)
# Interactively analyse graph
Inspect(SampleData,"StimRespPlot")

 # Build a time series. Average is taken from RespTrace between 0.8s and 1s for
 # each sweep and is plotted against the timestamp of the sweep. The graph
 # (ggplot) is returned directly.
BuildTimeSeriesPlot(SampleData,
   RespTrace = "I-mon",
   Time = c(0.8,1),
   fun = mean,
   ReturnPObject = FALSE)
}
\seealso{
\link[=Inspect]{Inspect}, \link[=Measure]{Measure}, \link[=GetPlot]{GetPlot}, \link[=GetPlotNames]{GetPlotNames}
}
