% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_load_nodes.R
\name{pm_load_nodes}
\alias{pm_load_nodes}
\title{pm_load_nodes}
\usage{
pm_load_nodes(con, nlevels, lvl_sizes, level)
}
\description{
function inherited from ephys2 (/github/tdanker/ephys2/). Undocumented.
}
