% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PMCollection.R
\docType{class}
\name{PMCollection-class}
\alias{PMCollection-class}
\alias{PMCollection}
\title{S4 class storing imported [Patch Master](https://www.heka.com/downloads/downloads_main.html#down_patchmaster)  Traces.}
\description{
This class stores imported [Patch Master](https://www.heka.com/downloads/downloads_main.html#down_patchmaster)  Traces. Currently only checked for time series. It is strictly validated to assure data consistency
}
\details{
\describe{
   \item{Series}{A list containing individual PMRecording objects of identical dimensisons, ProtocolName, RecMode and trace names}
   \item{Names}{Name identifier for the individual Series. Unique.}
   \item{Group}{Group identifier for the individual Series.}
   \item{MetaData}{A matrix with metatdata, e.g. calculated summay stats of each PMRecording object. One row per entry in Series allowed}
   \item{Plots}{A slot to store ggplots}
   \item{RecordingParams}{Stores the Recording parameters that must be identical for all entries in Series (ProtocolName, RecMode, Traces)}
 }
}
