% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMRecordingParams.R
\docType{class}
\name{PMRecordingParams-class}
\alias{PMRecordingParams-class}
\alias{PMRecordingParams}
\title{S4 class storing Recording parameters for impoted [Patch Master](https://www.heka.com/downloads/downloads_main.html#down_patchmaster)  Traces.}
\description{
This class stores imported [Patch Master](https://www.heka.com/downloads/downloads_main.html#down_patchmaster)  Traces. Currently only tested for time series. It is strictly validated to assure data consistency
}
