% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMTrace.R
\docType{class}
\name{PMSeries-class}
\alias{PMSeries-class}
\alias{PMSeries}
\title{S4 class storing imported [Patch Master](https://www.heka.com/downloads/downloads_main.html#down_patchmaster)  Traces.}
\description{
This class stores imported [Patch Master](https://www.heka.com/downloads/downloads_main.html#down_patchmaster)  Traces. Currently only checked for time series. It is strictly validated to assure data consistency
}
\details{

}
