% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PRecording.R
\docType{class}
\name{PRecording-class}
\alias{PRecording-class}
\alias{PRecording}
\title{S4 class storing imported ePhys traces.}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr
This class stores (usually sweep-oriented) electrophysiology data. Currently only import procedures for HEKA's PatchMaster .dat and Axon's .abd files are implemented. It has some strict validity checks implemented to assure data consistency.
}
\section{Slots}{

\describe{
\item{\code{Traces}}{Character vector containing names of the Traces (=Channels) imported from the dat file and any subsequently computed Trace. Computed Traces have the same dimensions as imported.}

\item{\code{Units}}{Character vector containing the SI units for data stored in corresponding Trace. Order as in Traces.}

\item{\code{TimeTrace}}{Numeric vector containing the time points of the recording.}

\item{\code{Sweeps}}{Ordered vector containing the names of the sweeps.}

\item{\code{SweepTimes}}{Numeric vector containing start times corresponding to Sweeps.}

\item{\code{Data}}{List of matrices. One list item per Trace. Matrix rows correspond to TimeTrace, columns to Sweeps.}

\item{\code{MetaData}}{Matrix that can contain additional per-sweep Metadata. Per-time Metadata can be stored in the Data slot using the \code{\link[=AddTrace]{AddTrace}} method.}

\item{\code{Plots}}{List that can contain any ggplot derived from the data. List item names that equal Traces are reserved.}

\item{\code{Fits}}{List that can contain fitted models.}

\item{\code{RecordingParams}}{An item of class \code{PRecordingParams} containing recording parameters for that trace.}
}}

\seealso{
\url{https://www.heka.com/downloads/downloads_main.html#down_patchmaster}
}
