% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PRecordingParams.R
\docType{class}
\name{PRecordingParams-class}
\alias{PRecordingParams-class}
\alias{PRecordingParams}
\title{S4 class storing Recording parameters inside \linkS4class{PRecording} and A \linkS4class{PCollection} objects}
\description{
Usually not required to be manipulated by the user
}
