% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.bundletree.R
\name{read.bundletree}
\alias{read.bundletree}
\title{read.bundletree}
\usage{
read.bundletree(myfile, bundlename = ".pul", con = NA)
}
\description{
function inherited from ephys2 (/github/tdanker/ephys2/). Undocumented.
}
