% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_id}
\alias{recode_id}
\title{Recode individual rows in a data frame.}
\usage{
recode_id(data, data_recode, colname_id)
}
\arguments{
\item{data}{A data frame.}

\item{data_recode}{The correspondance table.}

\item{colname_id}{The identification colname in both data and data_recode tables.}
}
\value{
A recoded data frame.
}
\description{
The data frame is recoded according to a correspondance table containing a row id.\cr
The correpondance table contains a least three columns : colname, value and an identification colname named with colname_id.
}
