% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recoder_factor}
\alias{recoder_factor}
\title{recoder_factor}
\usage{
recoder_factor(table, table_recodage, table_niveaux = NULL,
  source = NULL, filtre = NULL, creer_champs = FALSE)
}
\arguments{
\item{table}{La table à recoder.}

\item{table_recodage}{La table de recodage.}

\item{table_niveaux}{La table des niveaux.}

\item{source}{Nom de la source à filtrer dans la table \code{table_recodage}.}

\item{filtre}{La valeur de filtre.}

\item{creer_champs}{Créer les champs présents dans la table de recodage ou de niveaux mais absents dans la table à recoder.}
}
\value{
La table recodée
}
\description{
recoder_factor
}
