% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodage.R
\name{recoder_individu}
\alias{recoder_individu}
\title{Recoder les champs d'une table a partir d'un identifiant}
\usage{
recoder_individu(table, table_recodage, .champ_id = "identifiant")
}
\arguments{
\item{table}{La table à recoder.}

\item{table_recodage}{La table de recodage.}

\item{.champ_id}{Le nom du champ "identifiant".}
}
\value{
La table recodée
}
\description{
Recoder les champs d'une table à partir d'un identifiant.
}
