% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{RecParam,PMRecording-method}
\alias{RecParam,PMRecording-method}
\title{RecParam}
\usage{
\S4method{RecParam}{PMRecording}(object, param)
}
\arguments{
\item{object}{A PMRecordingor PMCollection object}

\item{param}{parameter to fetch, can be either of Rpip, RSeal, Urest, Cs, Rs}
}
\description{
get Recording parameters from PMRecording or PMCollection
}
