% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{remove_duplicate}
\alias{remove_duplicate}
\title{Remove duplicate rows from a table according a column name.}
\usage{
remove_duplicate(data, var)
}
\arguments{
\item{data}{A data frame.}

\item{var}{A variable for which duplicate rows are removed.}
}
\value{
A data frame with no duplicate for column name.
}
\description{
Remove duplicate rows from a table according a column name.
}
