% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename}
\alias{rename}
\title{Rename data frame column names according to a correspondance table.}
\usage{
rename(data, data_rename, drop = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{data_rename}{A correspondance table between old and new column names.}

\item{drop}{If \code{TRUE} then all the columns not present in data_rename are removed.}
}
\value{
A renamed data frame.
}
\description{
The correspondance table data_remane must at least contains tow columns : column for the old colnames and rename for the new colnames.
}
\examples{
dplyr::tibble(var1 = "a", old = "b") \%>\%
  patchr::rename(dplyr::tibble(column = "old", rename = "new"), drop = FALSE)

}
