% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{renommer_champs}
\alias{renommer_champs}
\title{Renomme les champs d'un data frame selon une table de correspondance}
\usage{
renommer_champs(table, table_rename, source = NULL, filtre = NULL,
  drop = TRUE)
}
\arguments{
\item{table}{Un data frame.}

\item{table_rename}{Une table de correspondance entre anciens et nouveaux noms de champ.}

\item{source}{Nom de la source à filtrer dans la table \code{table_rename}.}

\item{filtre}{La valeur de filtre.}

\item{drop}{\code{TRUE}: les champs non-rensignés dans la table de correspondance sont supprimés; \code{FALSE}: tous les champs sont conservés.}
}
\value{
Un data frame dont les champs sont renommés.
}
\description{
Renomme les champs d'un data frame selon une table de correspondance.
}
