% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{replace_empty_to_na}
\alias{replace_empty_to_na}
\title{Replace empty character with NA in a data frame.}
\usage{
replace_empty_to_na(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A data frame without empty character.
}
\description{
Replace empty character with NA in a data frame.
}
\examples{
data <- dplyr::tibble(
  champ_1 = "",
  champ_2 = ""
)
patchr::replace_empty_to_na(data)

}
