% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show}
\alias{show}
\title{Show}
\arguments{
\item{object}{An S4 object of type \link[=PRecording]{PRecording} or \link[=PCollection]{PCollection}}
}
\description{
Default method \code{show} for \link[=PRecording]{PRecording} and \link[=PCollection]{PCollection} objects.
}
