% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{str_normalise_colnames}
\alias{str_normalise_colnames}
\title{Normalise a string for a use as table column names.}
\usage{
str_normalise_colnames(string)
}
\arguments{
\item{string}{Input character vector.}
}
\value{
A character vector.
}
\description{
Normalise a string for a use as table column names.
}
\examples{
patchr::str_normalise_colnames(
  c("Type d'unité Sirus : entreprise profilée ou unité légale",
   "Nic du siège")
)

}
