% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetData.R
\name{SubsetData,PMRecording-method}
\alias{SubsetData,PMRecording-method}
\title{Subset a PMRecording object}
\usage{
\S4method{SubsetData}{PMRecording}(
  object,
  Traces = TraceNames(object),
  Sweeps = SweepNames(object),
  Time = range(getTimeTrace(object)),
  TimeExclusive = F,
  nowarnings = F
)
}
\arguments{
\item{object}{a \link[=PMRecording]{PMRecording} or \link[=PMCollection]{PMCollection} object}

\item{Traces, Sweeps}{List of Traces/Sweeps to keep}

\item{Time}{either a range of time points to keep, or, if \code{TimeExclusive} is \code{TRUE}, then two particular time points}

\item{TimeExclusive}{Keep only the two time points stated under Time, not the range}

\item{Series}{Subset by Series name. Only for \link[=PMCollection]{PMCollection}.}

\item{Group}{Subset by Group name. Only for \link[=PMCollection]{PMCollection}.}
}
\value{
A \link[=PMRecording]{PMRecording} object
}
\description{
This function subsets \link[=PMRecording]{PMRecording} or \link[=PMCollection]{PMCollection} objects by Trace, Sweep or Time
}
