% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetData.R
\docType{methods}
\name{SubsetData,PMSeries-method}
\alias{SubsetData,PMSeries-method}
\title{Subsetting PMSeries objects}
\usage{
\S4method{SubsetData}{PMSeries}(object, Traces = getTraces(object),
  Sweeps = getSweeps(object), Time = range(getTimeTrace(object)),
  TimeExclusive = F)
}
\arguments{
\item{object}{a PMSeries object}

\item{Traces, Sweeps}{List of Traces/Sweeps to keep}

\item{Time}{either a range of time points to keep, or, if @param TimeExclusive is TRUE, then two particular time points}

\item{TimeExclusive}{Keep only the two time points stated under Time, not the range}
}
\description{
This function subsets PMSeries objects by Trace, Sweep or Time
}
