% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetData.R
\docType{methods}
\name{SubsetData,PMTrace-method}
\alias{SubsetData,PMTrace-method}
\title{Subsetting PMTrace objects}
\usage{
\S4method{SubsetData}{PMTrace}(object, Channels = getChannels(object),
  Sweeps = getSweeps(object), Time = range(getTimeTrace(object)),
  TimeExclusive = F)
}
\arguments{
\item{object}{a PMTrace object}

\item{Channels, Sweeps}{List of Channels/Sweeps to keep}

\item{Time}{either a range of time points to keep, or, if @param TimeExclusive is TRUE, then two particular time points}

\item{TimeExclusive}{Keep only the two time points stated under Time, not the range}
}
\description{
This function subsets PMTrace objects by Channel, Sweep or Time
}
