% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetData.R
\name{SubsetData}
\alias{SubsetData}
\title{Subset a PRecording or PCollection object}
\usage{
SubsetData(
  X,
  Traces = getTraceNames(X),
  Sweeps = getSweepNames(X),
  Time = range(getTimeTrace(X)),
  Series = NULL,
  Group = NULL,
  TimeExclusive = F,
  nowarnings = F
)
}
\arguments{
\item{X}{a \link[=PRecording]{PRecording} or \link[=PCollection]{PCollection} Object}

\item{Traces, Sweeps}{List of Traces/Sweeps to keep}

\item{Time}{either a range of time points to keep, or, if \code{TimeExclusive} is \code{TRUE}, then two particular time points}

\item{Series}{Subset by Series name. Only for \link[=PCollection]{PCollection}.}

\item{Group}{Subset by Group name. Only for \link[=PCollection]{PCollection}.}

\item{TimeExclusive}{Keep only the two time points stated under Time, not the range}
}
\value{
A \link[=PRecording]{PRecording} X
}
\description{
This function subsets \link[=PRecording]{PRecording} or \link[=PCollection]{PCollection} objects by Trace, Sweep or Time
}
