% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_maj.R
\name{suppression_lignes_fin}
\alias{suppression_lignes_fin}
\title{Supprime les lignes d'une table selon les valeurs vides d'un champ ou une expression}
\usage{
suppression_lignes_fin(table, champs_filtre, ligne_fin = NULL)
}
\arguments{
\item{table}{Un data frame.}

\item{champs_filtre}{Un vecteur de nom de champ pour lequel les valeurs vides sont supprimées.}

\item{filtre_expression}{Une expression de type character supprimant des lignes du data frame.}
}
\value{
un data frame.
}
\description{
Supprime les lignes d'une table selon les valeurs vides d'un champ ou une expression.
}
