% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{supprimer_doublons_champ}
\alias{supprimer_doublons_champ}
\title{Supprime les doublons d'une table selon un champ}
\usage{
supprimer_doublons_champ(table, champ)
}
\arguments{
\item{table}{Un data frame.}

\item{champ}{Un champ pour lequel les doublons sont supprimés.}
}
\value{
un data frame.
}
\description{
Supprime les doublons d'une table selon un champ.
}
