% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{SweepNames,PMRecording-method}
\alias{SweepNames,PMRecording-method}
\title{SweepNames}
\usage{
\S4method{SweepNames}{PMRecording}(object)
}
\arguments{
\item{object}{A \link[=PMRecording]{PMRecording} object}
}
\description{
get list of Sweeps from a \link[=PMRecording]{PMRecording} object
}
