% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_maj.R
\name{to_date}
\alias{to_date}
\title{Transcode un vecteur vers un type date}
\usage{
to_date(vecteur, origin = "1899-12-30")
}
\arguments{
\item{vecteur}{Un vecteur.}
}
\value{
Un vecteur de type date
}
\description{
Transcode un vecteur vers un type date
}
