% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_maj.R
\name{to_integer}
\alias{to_integer}
\title{Transcode un vecteur vers un type integer}
\usage{
to_integer(vecteur)
}
\arguments{
\item{vecteur}{Un vecteur.}
}
\value{
Un vecteur de type integer
}
\description{
Transcode un vecteur vers un type integer
}
