% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcode.R
\name{to_numeric}
\alias{to_numeric}
\title{Transcode un vecteur vers un type numerique}
\usage{
to_numeric(vecteur)
}
\arguments{
\item{vecteur}{Un vecteur.}
}
\value{
Un vecteur de type numérique.
}
\description{
Transcode un vecteur vers un type numérique.
}
