% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{TraceNames,PMRecording-method}
\alias{TraceNames,PMRecording-method}
\title{TraceNames}
\usage{
\S4method{TraceNames}{PMRecording}(object)
}
\arguments{
\item{object}{A \link[=PMRecording]{PMRecording} or \link[=PMRecordingParams]{PMRecordingParams} object}
}
\description{
get list of Traces from a \link[=PMRecording]{PMRecording} or \link[=PMRecordingParams]{PMRecordingParams} object
}
