% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcode.R
\name{transcoder_champs}
\alias{transcoder_champs}
\title{Transcode les champs d'un data frame selon une table de correspondance}
\usage{
transcoder_champs(table, content_maj)
}
\arguments{
\item{table}{Un data frame.}

\item{content_maj}{Une table de correspondance entre les noms de champ et les classes attendus.}
}
\value{
Un data frame dont les champs sont transcodés.
}
\description{
Transcode les champs d'un data frame selon une table de correspondance.
}
