% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPMRecording.R
\name{addPRecording}
\alias{addPRecording}
\title{Adds PRecording to PCollection}
\usage{
addPRecording(
  object,
  PRecording,
  Names = if (is.list(PRecording)) {     lapply(PRecording, function(x)
    x@RecordingParams@Filename) } else {     PRecording@RecordingParams@Filename },
  Group
)
}
\arguments{
\item{object}{a \linkS4class{PRecording} object}

\item{PRecording}{a PRecording object, or list of PRecording objects}

\item{Names}{names for imported PRecording. Standard is variable name given in PRecording if single PRecording, i}

\item{Group}{Group variable or list thereof, of the same lenght as PRecording}
}
\value{
A \linkS4class{PCollection} object
}
\description{
Adds a \linkS4class{PRecording} to existing \linkS4class{PCollection}
}
