% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPRecording.R
\name{AddPRecording}
\alias{AddPRecording}
\title{Adds PRecording to PCollection}
\usage{
AddPRecording(
  X,
  PRecording,
  Names = if (is.list(PRecording)) {     lapply(PRecording, function(x)
    x@RecordingParams@Filename) } else {     PRecording@RecordingParams@Filename },
  Group = "Generic"
)
}
\arguments{
\item{X}{A \linkS4class{PRecording} object}

\item{PRecording}{a \var{PRecording}  object, or \var{list} of \var{PRecording} objects.}

\item{Names}{Name(s) for imported \var{PRecording}(s). Standard is the source file name as stored in the added a \var{PRecording}(s).}

\item{Group}{Group variable or list thereof, of the same length as \var{PRecording}}
}
\value{
A \linkS4class{PCollection} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr
Adds a \linkS4class{PRecording} to an existing \linkS4class{PCollection}
}
\examples{
# create a PCollection for testing
SampleCollection<-NewPCollection(SampleData,Names="A")
SampleCollection
# now add a PRecording
SampleCollection<-AddPRecording(SampleCollection, SampleData)
SampleCollection
}
