% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPMRecording.R
\name{AddPRecording}
\alias{AddPRecording}
\title{Adds PRecording to PCollection}
\usage{
AddPRecording(
  object,
  PRecording,
  Names = if (is.list(PRecording)) {     lapply(PRecording, function(x)
    x@RecordingParams@Filename) } else {     PRecording@RecordingParams@Filename },
  Group
)
}
\arguments{
\item{object}{A \linkS4class{PRecording} object}

\item{PRecording}{a \var{PRecording}  object, or \var{list} of \var{PRecording} objects.}

\item{Names}{Vames for imported a \var{PRecording}. Standard is File name stored in the added a \var{PRecording}(s).}

\item{Group}{Group variable or list thereof, of the same length as \var{PRecording}}
}
\value{
A \linkS4class{PCollection} object
}
\description{
Adds a \linkS4class{PRecording} to existing \linkS4class{PCollection}
}
