% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTrace.R
\name{addTrace,PMSeries-method}
\alias{addTrace,PMSeries-method}
\title{Add data (Trace) to PMSeries objects}
\usage{
\S4method{addTrace}{PMSeries}(object, Trace, Unit, Sweeps = colnames(mtx), isOrig = F, mtx)
}
\arguments{
\item{object}{A PMSeries object}

\item{Trace}{Name of the new Trace}

\item{Sweeps}{Names of the sweeps added. Must be the same as sweep names in \code{object}. Data will be sorted accoding to order of Sweeps in \code{object} Default is \code{colnames(object)}}

\item{mtx}{Any object convertible into a matrix, that has the same dimension as data in the Data slot of \code{object}}
}
\value{
A matrix or \link[=PMSeries]{PMSeries} object
}
\description{
This function adds a new Trace (with data) from any object convertible into a matrix to a PMSeries object
}
