% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\name{apply,PMRecording-method}
\alias{apply,PMRecording-method}
\title{apply function for PMRecording objects}
\usage{
\S4method{apply}{PMRecording}(X, MARGIN, FUN, ReturnPMRecording = F)
}
\arguments{
\item{X}{a PMRecording object}

\item{MARGIN}{a vector giving the subscripts which the function will be applied along. Understands "Time", "Sweep","Trace", or 1-3 resp.}

\item{FUN}{the function to be applied}

\item{ReturnPMRecording}{whether to return results as a PMRecording with an additional, computed trace. Default is \var{FALSE}, then returns a matrix.}
}
\value{
A matrix or \link[=PMRecording]{PMRecording} object
}
\description{
This is the PatchMastR analog to "apply"
}
