% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.R
\name{as.data.frame,PRecording-method}
\alias{as.data.frame,PRecording-method}
\title{Converts PRecording or PCollection into a long data frame}
\usage{
\S4method{as.data.frame}{PRecording}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{a \var{PRecording} or \var{PCollection} object.}
}
\value{
A \var{data.frame}.
}
\description{
Converts \linkS4class{PRecording} or \linkS4class{PCollection}  into a data frame in a long representation, analog to tidyR's gather or
}
\seealso{
\link[base::as.data.frame]{as.data.frame()},\link[tidyr::gather]{gather()},\link[tidyr::pivot_longer]{pivot_longer()}
}
