% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.R
\docType{methods}
\name{as.data.frame}
\alias{as.data.frame}
\title{toLong}
\usage{
\S4method{as.data.frame}{PMSeries}(x, row.names = NULL,
  optional = FALSE, ...)
}
\arguments{
\item{x}{a PMSeries x}
}
\description{
converts PMSeries x into a data frame in a long representation, analog to tidyR's gather
}
