% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.R
\name{as.data.frame}
\alias{as.data.frame}
\alias{as.data.frame,PRecording-method}
\title{Converts PRecording or PCollection into a long data frame}
\usage{
\S4method{as.data.frame}{PRecording}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A \var{PRecording} or \var{PCollection} object.}

\item{row.names, optional, ...}{Unused.}
}
\value{
A \code{data.frame}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr
Converts \linkS4class{PRecording} or \linkS4class{PCollection} into a data frame in a long representation, analog to tidyR's \link[tidyr:gather]{gather()}/\link[tidyr:pivot_longer]{pivot_longer()}.
}
\section{Methods (by class)}{
\itemize{
\item \code{PRecording}: Method for PRecording
}}

\examples{
data(PRecording)
# Subset SampleData to only keep three time points and convert into data.frame
as.data.frame(GetData(SampleData,Time = c(0.4,0.8,1.2),TimeExclusive = T))
}
\seealso{
\link[base:as.data.frame]{as.data.frame()}, \link[tidyr:gather]{gather()}, \link[tidyr:pivot_longer]{pivot_longer()}
}
