% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as_date}
\alias{as_date}
\title{Transcode a vector into date.}
\usage{
as_date(x, origin = "1899-12-30")
}
\arguments{
\item{x}{A vector.}
}
\value{
A date vector.
}
\description{
Transcode a vector into date.
}
\keyword{internal}
