% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{as_factor}
\alias{as_factor}
\title{as_factor}
\usage{
as_factor(champ_factor, table_niveaux = NULL)
}
\arguments{
\item{champ_factor}{\dots}

\item{table_niveaux}{\dots}
}
\description{
as_factor
}
\keyword{internal}
