% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as_factor}
\alias{as_factor}
\title{Transcode a character vector into an factor.}
\usage{
as_factor(string, data_levels = NULL)
}
\arguments{
\item{string}{A character vector.}

\item{data_levels}{An optional correpondance table between column names and ordered levels.}
}
\value{
A ordered leveled factor.
}
\description{
An optional table data_levels contains ordered levels associated with column names. It contains at least three colmuns : column, order, level.
}
\keyword{internal}
