% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildTimeSeriesPlot.R
\docType{methods}
\name{BuildTimeSeriesPlot,PMSeries-method}
\alias{BuildTimeSeriesPlot,PMSeries-method}
\title{#' Basic ggplots for time series data}
\usage{
\S4method{BuildTimeSeriesPlot}{PMSeries}(object)
}
\arguments{
\item{object}{a PMSeries object}
}
\description{
This function creats a basic visualization for each Trace in the object. These will be stored in the Plots slot
}
