% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_IV.R
\docType{methods}
\name{Calculate_IV,PMSeries-method}
\alias{Calculate_IV,PMSeries-method}
\title{Averaging PMSeries objects}
\usage{
\S4method{Calculate_IV}{PMSeries}(object, X_FROM, X_TO,
  ReturnPMTRace = T, I_Trace = "I-mon", V_Trace = "V-mon",
  x_D_FROM = NA, X_D_TO = NA)
}
\arguments{
\item{object}{a PMSeries object}

\item{X_FROM, X_TO, x_D_FROM, X_D_TO}{Time points to perform averaging for IV prodcution}

\item{ReturnPMTRace}{whether to return results as a PMSeries with an additional, computed trace}

\item{I_Trace, V_Trace}{Name of the traces containig I and V}
}
\value{
a matrix or PMSeries with IV data and ggplot stored in the MetaData and Plot slot, resp.
}
\description{
This function averages PMSeries objects by Trace, Sweep or Time
}
