% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenientScales.R
\docType{methods}
\name{convenientScales}
\alias{convenientScales}
\title{convenientScales}
\usage{
\S4method{convenientScales}{PMTrace}(object)
}
\arguments{
\item{object}{a PMTrace object}
}
\description{
This function converts scaling of a PMTrace object and adjusts/adds SI prefixes
}
