% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenientScales.R
\name{convenientScales}
\alias{convenientScales}
\title{Makes PRecording object scales convenient.}
\usage{
convenientScales(object)
}
\arguments{
\item{object}{a \link[=PRecording]{PRecording} object}
}
\value{
a \link[=PRecording]{PRecording} object
}
\description{
This function converts scaling of a \link[=PRecording]{PRecording} or a \link[=PCollection]{PCollection} object and adjusts/adds SI prefixes
}
