% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenientScales.R
\name{ConvenientScales}
\alias{ConvenientScales}
\title{Makes PRecording object scales convenient.}
\usage{
ConvenientScales(X)
}
\arguments{
\item{X}{A \linkS4class{PRecording} or \linkS4class{PCollection} object}
}
\value{
A \linkS4class{PRecording} or \linkS4class{PCollection} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr
This function converts scaling of a \linkS4class{PRecording} or \linkS4class{PCollection} object and adds SI prefixes. Caution: Currently no check is performed if a unit already carriesa SI prefix.
}
\seealso{
\link[sitools:f2si]{sitools::f2si()}
}
