% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurrentDensity.R
\name{CurrentDensity}
\alias{CurrentDensity}
\alias{CurrentDensity,PRecording-method}
\alias{CurrentDensity,PCollection-method}
\title{Normalizes current traces to cell capacity}
\usage{
CurrentDensity(X, Trace = "I.mon")

\S4method{CurrentDensity}{PRecording}(X, Trace = "I.mon")

\S4method{CurrentDensity}{PCollection}(X, Trace = "I.mon")
}
\arguments{
\item{X}{A \linkS4class{PRecording} or \linkS4class{PCollection} object}

\item{Trace}{Name of trace containing current data to normalize to cell capacity.}
}
\value{
A \linkS4class{PRecording} or \linkS4class{PCollection} object.
}
\description{
This function normalizes current traces to cell capacity and adds it as a new Trace
}
\section{Functions}{
\itemize{
\item \code{CurrentDensity,PRecording-method}: Method for PRecording

\item \code{CurrentDensity,PCollection-method}: Method for PCollection
}}

