% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurrentDensity.R
\name{CurrentDensity}
\alias{CurrentDensity}
\title{CurrentDensity}
\usage{
\S4method{CurrentDensity}{PMSeries}(object, trace = "I.mon")
}
\arguments{
\item{object}{a PMSeries object}
}
\description{
This function normalizes current traces to cell capacity
}
