% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatchR-data.R
\docType{data}
\name{.SampleData}
\alias{.SampleData}
\alias{PRecording}
\title{IV Protocol data}
\format{
An object of class \code{"PRecording"}; see \code{\link[PRecording]{PRecording}}.
}
\usage{
data(PRecording)
}
\description{
This dataset contains data from three traces from an IV recording protocol.
}
\examples{
rec<-Load(data(PRecording))
rec<-BuildRecordingPlot(rec)
Inspect(rec, what = "I-mon")
}
\keyword{datasets}
