% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{downsample,PMSeries-method}
\alias{downsample,PMSeries-method}
\title{Donwsample a PMSeries object}
\usage{
\S4method{downsample}{PMSeries}(object, npnts.out = NULL, sample.rate = NULL)
}
\arguments{
\item{object}{a \link[=PMSeries]{PMSeries} object}

\item{npnts.out}{Number of points in output time axis, ignored if NULL}

\item{sample.rate}{New sampling rate in 1/[object@TimeUnit]}
}
\value{
A matrix or \link[=PMSeries]{PMSeries} object
}
\description{
Downsample \link[=PMSeries]{PMSeries} object along the time axis
}
