% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get-methods.R, R/GetData.R
\name{Get}
\alias{Get}
\alias{GetSweepNames}
\alias{GetTraceNames}
\alias{GetTimeTrace}
\alias{GetSweepTimes}
\alias{GetCSlow}
\alias{GetRecParam}
\alias{GetGroupMembers}
\alias{GetMetaData}
\alias{GetData}
\title{Accession methods}
\usage{
GetSweepNames(X)

GetTraceNames(X)

GetTimeTrace(X)

GetSweepTimes(X)

GetCSlow(X)

GetRecParam(X, which)

GetGroupMembers(X, which)

GetMetaData(X, which)

GetData(
  X,
  Traces = GetTraceNames(X),
  Sweeps = GetSweepNames(X),
  Time = range(GetTimeTrace(X)),
  Series = NULL,
  Group = NULL,
  TimeExclusive = F,
  nowarnings = F
)
}
\arguments{
\item{X}{A \linkS4class{PRecording} or \linkS4class{PCollection} object}

\item{which}{A name of a valid Group in a \linkS4class{PCollection} (for \code{GetGroupMembers}). \cr A name or a vector of names of slot(s) in RecordingParams (for \code{GetRecParam}). \cr A name or a vector of names of a column in the MetaData slot (for \code{GetMetaData}).}

\item{Traces, Sweeps}{List of Traces/Sweeps to keep}

\item{Time}{either a range of time points to keep, or, if \code{TimeExclusive} is \code{TRUE}, then two particular time points}

\item{Series}{Subset by Series name. Only for  \linkS4class{PCollection} .}

\item{Group}{Subset by Group name. Only for  \linkS4class{PCollection} .}

\item{TimeExclusive}{Keep only the two time points stated under Time, not the range}

\item{nowarnings}{Supress warning messages.}
}
\value{
A numeric vector.

For \code{GetMetaData} a \code{vector} of type \code{numeric}  or a \code{matrix}.

A \linkS4class{PRecording} or \linkS4class{PCollection} object (for \code{GetData})
}
\description{
These methods are used to access information from \linkS4class{PRecording} and/or \linkS4class{PCollection} objects
}
\details{
These methods can be used to access information stored in \linkS4class{PRecording} and/or \linkS4class{PCollection} objects. \cr \cr
}
\section{Functions}{
\itemize{
\item \code{GetSweepNames}: Returns the Sweep names.

\item \code{GetTraceNames}: Return the Trace names.

\item \code{GetTimeTrace}: Returns a vector containing the times at which the individual sweeps were recorded.

\item \code{GetSweepTimes}: Returns a vector containing the times at which the individual sweeps were recorded.

\item \code{GetCSlow}: This is a special case of \code{GetRecParam} and returns the C-slow value.

\item \code{GetRecParam}: Returns any value stored in the RecordingParams slot.

\item \code{GetGroupMembers}: Can be used on \linkS4class{PCollection} objects only and returns the names of all PRecordings belonging to the named Group.

\item \code{GetMetaData}: Returns one or more columns from the MetaData Slot.

\item \code{GetData}: This function subsets objects by \var{Trace}, \var{Sweep} or \var{Time}. For  \linkS4class{PCollection} additionally by \var{Series} or \var{Group}
}}

\examples{
\dontrun{
GetRecParam(recording,c("Cs","Rs"))
}
}
