% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get}
\alias{get}
\alias{getSweepNames}
\alias{getTraceNames}
\alias{getTimeTrace}
\alias{getCSlow}
\alias{getRecParam}
\alias{getGroupMembers}
\alias{getMetaData}
\title{Accession methods}
\usage{
getSweepNames(object)

getTraceNames(object)

getTimeTrace(object)

getCSlow(object)

getRecParam(object, which)

getGroupMembers(object, which)

getMetaData(object, which)
}
\arguments{
\item{object}{A \linkS4class{PMRecording} or \linkS4class{PMCollection}  object}

\item{which}{A name of a valid Group in a \linkS4class{PMCollection} (for \code{getGroupMembers}). \cr A name or a vector of names of slot(s) in RecordingParams (for \code{getRecParam}). \cr A name or a vector of names of a column in the MetaData slot (for \code{getMetaData}).}
}
\value{
A numeric vector or a matrix (for \code{getMetaData}).
}
\description{
These methods are used to access information from \linkS4class{PMRecording} and/or \linkS4class{PMCollection} objects
}
\details{
These methods can be used to access information stored in \linkS4class{PMRecording} and/or \linkS4class{PMCollection} objects. \cr \cr
}
\section{Functions}{
\itemize{
\item \code{getSweepNames}: Can be used on \linkS4class{PMRecording} and \linkS4class{PMCollection} objects and return the Sweep names.

\item \code{getTraceNames}: Can be used on \linkS4class{PMRecording} and \linkS4class{PMCollection} objects and return the Trace names.

\item \code{getTimeTrace}: Can be used on \linkS4class{PMRecording} and \linkS4class{PMCollection} objects and returns a vector containing the times at which the individual sweeps were recorded.

\item \code{getCSlow}: is a special case of \code{getRecParam} and can be used on \linkS4class{PMRecording} and \linkS4class{PMCollection} objects. It returns the C-slow value.

\item \code{getRecParam}: Can be used on \linkS4class{PMRecording} and \linkS4class{PMCollection} objects and returns any value stored in the RecordingParams slot.

\item \code{getGroupMembers}: get Can be used on \linkS4class{PMCollection} objects and returns the names of all PMRecordings belonging to the named Group.

\item \code{getMetaData}: Can be used on \linkS4class{PMRecording} and \linkS4class{PMCollection} objects and returns one or more columns from the MetaData Slot.
}}

\examples{
getRecParam(recording,c("Cs","Rs"))
}
