% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetData.R, R/get.R
\name{getData}
\alias{getData}
\alias{get}
\alias{getSweepNames}
\alias{getTraceNames}
\alias{getTimeTrace}
\alias{getSweepTimes}
\alias{getCSlow}
\alias{getRecParam}
\alias{getGroupMembers}
\alias{getMetaData}
\title{Accession methods}
\usage{
getData(
  X,
  Traces = getTraceNames(X),
  Sweeps = getSweepNames(X),
  Time = range(getTimeTrace(X)),
  Series = NULL,
  Group = NULL,
  TimeExclusive = F,
  nowarnings = F
)

getSweepNames(X)

getTraceNames(X)

getTimeTrace(X)

getSweepTimes(X)

getCSlow(X)

getRecParam(X, which)

getGroupMembers(X, which)

getMetaData(X, which)
}
\arguments{
\item{X}{A \linkS4class{PRecording} or \linkS4class{PCollection} object}

\item{Traces, Sweeps}{List of Traces/Sweeps to keep}

\item{Time}{either a range of time points to keep, or, if \code{TimeExclusive} is \code{TRUE}, then two particular time points}

\item{Series}{Subset by Series name. Only for  \linkS4class{PCollection} .}

\item{Group}{Subset by Group name. Only for  \linkS4class{PCollection} .}

\item{TimeExclusive}{Keep only the two time points stated under Time, not the range}

\item{which}{A name of a valid Group in a \linkS4class{PCollection} (for \code{getGroupMembers}). \cr A name or a vector of names of slot(s) in RecordingParams (for \code{getRecParam}). \cr A name or a vector of names of a column in the MetaData slot (for \code{getMetaData}).}
}
\value{
A \linkS4class{PRecording} or \linkS4class{PCollection} object

A numeric vector or a matrix (for \code{getMetaData}).
}
\description{
These methods are used to access information from \linkS4class{PRecording} and/or \linkS4class{PCollection} objects
}
\details{
These methods can be used to access information stored in \linkS4class{PRecording} and/or \linkS4class{PCollection} objects. \cr \cr
}
\section{Functions}{
\itemize{
\item \code{getData}: This function subsets \linkS4class{PRecording} or \linkS4class{PCollection} objects by \var{Trace}, \var{Sweep} or \var{Time}. For  \linkS4class{PCollection} additionally by \var{Series} or \var{Group}

\item \code{getSweepNames}: Can be used on \linkS4class{PRecording} and \linkS4class{PCollection} objects and return the Sweep names.

\item \code{getTraceNames}: Can be used on \linkS4class{PRecording} and \linkS4class{PCollection} objects and return the Trace names.

\item \code{getTimeTrace}: Can be used on \linkS4class{PRecording} and \linkS4class{PCollection} objects and returns a vector containing the times at which the individual sweeps were recorded.

\item \code{getSweepTimes}: Can be used on \linkS4class{PRecording} and \linkS4class{PCollection} objects and returns a vector containing the times at which the individual sweeps were recorded.

\item \code{getCSlow}: is a special case of \code{getRecParam} and can be used on \linkS4class{PRecording} and \linkS4class{PCollection} objects. It returns the C-slow value.

\item \code{getRecParam}: Can be used on \linkS4class{PRecording} and \linkS4class{PCollection} objects and returns any value stored in the RecordingParams slot.

\item \code{getGroupMembers}: get Can be used on \linkS4class{PCollection} objects and returns the names of all PRecordings belonging to the named Group.

\item \code{getMetaData}: Can be used on \linkS4class{PRecording} and \linkS4class{PCollection} objects and returns one or more columns from the MetaData Slot.
}}

\examples{
getRecParam(recording,c("Cs","Rs"))
}
