% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get-methods.R
\name{Get}
\alias{Get}
\alias{GetSweepNames}
\alias{GetTraceNames}
\alias{GetTimeTrace}
\alias{GetSweepTimes}
\alias{GetCSlow}
\alias{GetRecParam}
\alias{GetGroupMembers}
\alias{GetMetaData}
\title{Accession methods}
\usage{
GetSweepNames(X)

GetTraceNames(X)

GetTimeTrace(X)

GetSweepTimes(X)

GetCSlow(X)

GetRecParam(X, which)

GetGroupMembers(X, which)

GetMetaData(X, which)
}
\arguments{
\item{X}{A \linkS4class{PRecording} or \linkS4class{PCollection} object}

\item{which}{A name of a valid Group in a \linkS4class{PCollection} (for \code{GetGroupMembers}). \cr A name or a vector of names of slot(s) in RecordingParams (for \code{GetRecParam}). \cr A name or a vector of names of a column in the MetaData slot (for \code{GetMetaData}).}
}
\value{
A numeric vector.

For \code{GetMetaData} a \code{vector} of type \code{numeric}  or a \code{matrix}.
}
\description{
These methods are used to access information from \linkS4class{PRecording} and/or \linkS4class{PCollection} objects
}
\details{
These methods can be used to access information stored in \linkS4class{PRecording} and/or \linkS4class{PCollection} objects. \cr \cr
}
\section{Functions}{
\itemize{
\item \code{GetSweepNames}: Returns the Sweep names.

\item \code{GetTraceNames}: Return the Trace names.

\item \code{GetTimeTrace}: Returns a vector containing the times at which the individual sweeps were recorded.

\item \code{GetSweepTimes}: Returns a vector containing the times at which the individual sweeps were recorded.

\item \code{GetCSlow}: This is a special case of \code{GetRecParam} and returns the C-slow value.

\item \code{GetRecParam}: Returns any value stored in the RecordingParams slot.

\item \code{GetGroupMembers}: Can be used on \linkS4class{PCollection} objects only and returns the names of all PRecordings belonging to the named Group.

\item \code{GetMetaData}: Returns one or more columns from the MetaData Slot.
}}

\examples{
\dontrun{
GetRecParam(recording,c("Cs","Rs"))
}
}
