% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get-methods.R, R/GetData.R
\name{Get}
\alias{Get}
\alias{GetSweepNames}
\alias{GetTraceNames}
\alias{GetTimeTrace}
\alias{GetSweepTimes}
\alias{GetCSlow}
\alias{GetRecParam}
\alias{GetGroupMembers}
\alias{GetGroupNames}
\alias{GetGroups}
\alias{GetRecordingNames}
\alias{GetMetaData}
\alias{GetPlotNames}
\alias{GetPlot}
\alias{GetData}
\alias{Subset}
\title{Accession methods}
\usage{
GetSweepNames(X)

GetTraceNames(X)

GetTimeTrace(X)

GetSweepTimes(X)

GetCSlow(X)

GetRecParam(X, which)

GetGroupMembers(X, which)

GetGroupNames(X)

GetGroups(X)

GetRecordingNames(X)

GetMetaData(X, which = colnames(X@MetaData))

GetPlotNames(X)

GetPlot(X, which = colnames(X@MetaData))

GetData(
  X,
  Traces = GetTraceNames(X),
  Sweeps = GetSweepNames(X),
  Time = range(GetTimeTrace(X)),
  Recordings = NULL,
  Group = NULL,
  TimeExclusive = F,
  nowarnings = F
)

Subset(
  X,
  Traces = GetTraceNames(X),
  Sweeps = GetSweepNames(X),
  Time = range(GetTimeTrace(X)),
  Recordings = NULL,
  Group = NULL,
  TimeExclusive = F,
  nowarnings = F
)
}
\arguments{
\item{X}{A \linkS4class{PRecording} or \linkS4class{PCollection} object}

\item{which}{\enumerate{
\item A name of a valid Group in a \linkS4class{PCollection} (for \code{GetGroupMembers}). \cr 2) A name or a vector of names of slot(s) in RecordingParams (for \code{GetRecParam}). \cr or 3) A name or a vector of names of a column in the MetaData slot (for \code{GetMetaData}).
}}

\item{Traces}{List of traces/channels to keep}

\item{Sweeps}{List of sweeps to keep}

\item{Time}{either a range of time points to keep or two particular time points}

\item{Recordings}{Subset by series/recordings. Understands names (= file names of the recordings) or indices or by logical indexing. Only for \linkS4class{PCollection} .}

\item{Group}{Subset by Group name. Only for  \linkS4class{PCollection} .}

\item{TimeExclusive}{Keep only the two time points stated under \code{Time}, not the range}

\item{nowarnings}{Supress warning messages.}
}
\value{
A numeric vector.

For \code{GetMetaData} a \code{vector} of type \code{numeric}  or a \code{matrix}.

For \code{GetPlotNames} a \code{vector} of type \code{character}.

For \code{GetPlot} a \link[=ggplot2]{ggplot}.

For \code{GetData} A \linkS4class{PRecording} or \linkS4class{PCollection} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr
These methods are used to access information from \linkS4class{PRecording} and/or \linkS4class{PCollection} objects
}
\details{
These methods can be used to access information stored in \linkS4class{PRecording} and/or \linkS4class{PCollection} objects. \cr \cr
}
\section{Functions}{
\itemize{
\item \code{GetSweepNames}: Returns the Sweep names.

\item \code{GetTraceNames}: Return the Trace names.

\item \code{GetTimeTrace}: Returns a vector containing the times at which the individual sweeps were recorded.

\item \code{GetSweepTimes}: Returns a vector containing the times at which the individual sweeps were recorded.

\item \code{GetCSlow}: This is a special case of \code{GetRecParam} and returns the C-slow value.

\item \code{GetRecParam}: Returns any value stored in the RecordingParams slot.

\item \code{GetGroupMembers}: Can be used on \linkS4class{PCollection} objects only and returns the names of all PRecordings belonging to the named Group.

\item \code{GetGroupNames}: Can be used on \linkS4class{PCollection} objects only and returns the names of all groups.

\item \code{GetGroups}: Can be used on \linkS4class{PCollection} objects only and returns the groups.

\item \code{GetRecordingNames}: Can be used on \linkS4class{PCollection} objects only and returns the names of all recordings.

\item \code{GetMetaData}: Returns one or more columns from the MetaData Slot.

\item \code{GetPlotNames}: Returns the names of the plots stored in the Plots Slot.

\item \code{GetPlot}: Returns a plot from the Plots slot.

\item \code{GetData}: This method subsets objects by \var{Trace}, \var{Sweep} or \var{Time}. For  \linkS4class{PCollection} additionally by \var{Recordings} or \var{Group}

\item \code{Subset}: Subset is an alias of Getdata
}}

