% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getSweeps,PMSeries-method}
\alias{getSweeps,PMSeries-method}
\title{getSweeps}
\usage{
\S4method{getSweeps}{PMSeries}(object)
}
\arguments{
\item{object}{A \link[=PMSeries]{PMSeries} object}
}
\description{
get list of Sweeps from a \link[=PMSeries]{PMSeries} object
}
