% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getTimeTrace,PMSeries-method}
\alias{getTimeTrace,PMSeries-method}
\title{getTimeTrace}
\usage{
\S4method{getTimeTrace}{PMSeries}(object)
}
\arguments{
\item{object}{A \link[=PMSeries]{PMSeries}  object}
}
\description{
get Time trace from a \link[=PMSeries]{PMSeries}  object
}
