% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportCollection.R
\name{ImportCollection}
\alias{ImportCollection}
\title{Imports a collection of PatchClamp recordings}
\usage{
ImportCollection(
  parent.dir = getwd(),
  filelist = NULL,
  traces = c(1, 2),
  type = "PatchMaster",
  encoding = getOption("encoding")
)
}
\arguments{
\item{parent.dir}{parent directory where to search for FileNames from \code{filelist}}

\item{filelist}{a data.frame with four columns: "FileName","Experiment","Series","Group". FileName can be absolute or relative to parent.dir. "Experiment","Series" as defined in PatchMaster file. Groups can be assigned via the Group variable}

\item{traces}{which traces to import. default is \code{c(1,2)}}

\item{type}{File Type. Currently not implemented. Only works with PatchMatster files.}

\item{encoding}{file encoding to use, default is \code{getOption("encoding")}}
}
\value{
A \link[=PCollection]{PCollection} object
}
\description{
Imports a collection of PatchClamp recordings. I e those that be long to a the same experimental series.
}
