% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportPMRecording.R
\name{ImportPRecording}
\alias{ImportPRecording}
\title{Import Series from Patch Master *.dat file into a PRecording object}
\usage{
ImportPRecording(
  filename,
  experiment = 1,
  series = 1,
  traces = c(1, 2),
  encoding = getOption("encoding")
)
}
\arguments{
\item{filename}{path to a [Patch Master](https://www.heka.com/downloads/downloads_main.html#down_patchmaster) *.dat file.}

\item{experiment}{Index of Experiment to import (1st order hierarchy in PM dat file).}

\item{series}{Index of Series to import (2nd order hierarchy in PM dat file).}

\item{traces}{Traces to import (indices). Must be vector of numeric of length > 0.}

\item{encoding}{file encoding to use, default is \code{getOption("encoding")}}
}
\value{
A \link[=PRecording]{PRecording} object
}
\description{
This function imports Series from Patch Master *.dat files and creates PRecording objects
}
