% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectTimeSeries.R
\docType{methods}
\name{InspectTimeSeries}
\alias{InspectTimeSeries}
\title{InspectTimeSeries}
\usage{
\S4method{InspectTimeSeries}{PMSeries}(object, Trace = "I.mon")
}
\arguments{
\item{object}{a PMSeries object}

\item{Trace}{Trace to plot}
}
\description{
This function creates an interactive graph from a plot generated using BuildTimeSeriesPlot
}
