% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapply.R
\name{lapply,PMCollection-method}
\alias{lapply,PMCollection-method}
\title{apply function for PMCollection objects}
\usage{
\S4method{lapply}{PMCollection}(X, FUN, ReturnPMCollection = F)
}
\arguments{
\item{X}{a PMCollection object}

\item{FUN}{the function to be applied}

\item{ReturnPMCollection}{whether to return results as a PMCollection. Default is \var{FALSE}, then returns a list}
}
\value{
A \link[=PMCollection]{PMCollection} object
}
\description{
This is the PatchMastR analog to "lapply" for PMCollection objects
}
