% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load.R
\name{Load}
\alias{Load}
\title{(OK) Load HDF5 files containing PRecording or PCollection}
\usage{
Load(filename)
}
\arguments{
\item{filename}{Path to the file to be loaded.}
}
\value{
A\linkS4class{PCollection} or \linkS4class{PCollection} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr
This function loads HDF5 files containing \linkS4class{PCollection} or \linkS4class{PCollection} data.
}
\seealso{
Save
}
