% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatchMasteR.R
\docType{package}
\name{PatchMasteR}
\alias{PatchMasteR}
\title{PatchMasteR: An environment for handling and analzing Patch Clamp measuremnts recorded with HEKA Patchmaster}
\description{
Core component of PatchMasteR is the PMRecording class, wich stores a series imported from a Patchmaster *.dat file. Several integrety checks are implemented into the PMRecording class to assure data integrity is maintained.
Multiple recordings (PMRecording) objects can be stored and processed within the PMCollection class.
}
\section{Import and creation functions}{

* \link[=ImportPMRecording]{ImportPMRecording} (for \link[=PMRecording]{PMRecording} objects) \cr
* \link[=addTrace]{addTrace} (for \link[=PMRecording]{PMRecording} objects) \cr
* \link[=newPMCollection]{newPMCollection} (for \link[=PMCollection]{PMCollection} objects) \cr
* \link[=addPMRecording]{addPMRecording} (for \link[=PMCollection]{PMCollection} objects) \cr
* \link[=dropPMRecording]{addPMRecording} (for \link[=PMCollection]{PMCollection} objects) \cr
}

\section{Accession functions for \link[=PMRecording]{PMRecording} objects}{

* \link[=getCs]{getCs}
* \link[=SweepNames]{SweepNames}
* \link[=getTimeTrace]{getTimeTrace}
* \link[=TraceNames]{TraceNames}
* \link[=getMetaData]{getMetaData}
* \link[=RecParam]{RecParam}
* \link[=as.data.frame]{as.data.frame}
}

\section{Accession functions for \link[=PMCollection]{PMCollection} objects}{

* \link[=as.data.frame]{as.data.frame}
* \link[=getMetaData]{getMetaData}
* \link[=RecParam]{RecParam}
}

\section{Visuaization functions}{

* \link[=BuildSeriesPlot]{BuildSeriesPlot} (for \link[=PMRecording]{PMRecording} objects)
* \link[=BuildQCPlot]{BuildSeriesPlot} (for \link[=PMCollection]{PMCollection} objects)
* \link[=Inspect]{Inspect} (for \link[=PMRecording]{PMRecording} and \link[=PMCollection]{PMCollection} oojects)
*visualize time series
}

\section{Math, conversion and subsampling functions}{

* \link[=CurrentDensity]{CurrentDensity} (for \link[=PMRecording]{PMRecording} objects)
* \link[=SubsetData]{SubsetData}
* \link[=CalculateIV]{CalculateIV} CHECK!!!
* \link[=apply]{apply} for \link[=PMRecording]{PMRecording} and \link[=lapply]{lapply} for \link[=PMCollection]{PMCollection}
}

\author{
\link[https://www.uni-marburg.de/en/fb20/departments/physiology/research/dominik-oliver-lab/research2/retinal-physiology-and-gene-therapy]{Moritz Lindner}
}
