% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackDesc.R
\docType{package}
\name{PatchMasteR}
\alias{PatchMasteR}
\alias{PatchMasteR-package}
\title{PatchMasteR: An environment for handling and analzing Patch Clamp measuremnts recorded with HEKA Patchmaster}
\description{
Core component of PatchMasteR is the Class PMTrace, wich Stores a Series imported from a Patchmaster *.dat file
}
\section{Import functions}{

ImportPMSeries
}

\section{Accession functions}{

getCs \n
getSweeps \n
getTimeTrace \n
getTraces \n
as.data.frame \n
}

\section{Visuaization functions}{

BuildTimeSeriesPlot
InspectTimeSeries
}

\section{Math, conversion and subsampling functions}{

CurrentDensity
SubsetData
CalculateIV
apply
}

