% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatchMasteR-package.R
\docType{package}
\name{PatchMasteR}
\alias{PatchMasteR}
\alias{PatchMasteR-package}
\title{PatchMasteR: An environment for working with Traces imorted form HEKA's PatchMaster}
\description{
An environment for working with Traces imorted form HEKA's PatchMaster
}
\keyword{internal}
