% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PCollection.R
\docType{class}
\name{PCollection-class}
\alias{PCollection-class}
\alias{PCollection}
\title{S4 class storing a collection of ePhys Treaces}
\description{
This class stores a collection of \linkS4class{PRecording}s in a single object. Facilitates identical processing of related of recordings.
}
\section{Slots}{

\describe{
\item{\code{Series}}{A list containing individual \linkS4class{PRecording} objects of identical dimensions, \var{ProtocolName}, \var{RecMode} and \var{TraceNames}.}

\item{\code{Names}}{A \var{vector} of type \code{character} as unique identifier for the individual PRecording objects stored in the collection.}

\item{\code{Group}}{A \var{vector} of type \code{logical} as Group identifier for the individual Series.}

\item{\code{MetaData}}{A \var{matrix} with each row corresponding to a \linkS4class{PRecording} stored in the \linkS4class{PCollection}. Column names must be unique.}

\item{\code{Plots}}{A slot to store ggplots}

\item{\code{.MetaDataFx}}{A list fo functions called to under \link[=AddMetaData]{AddMetaData()}}

\item{\code{RecordingParams}}{Stores the Recording parameters that must be identical for all entries in Series \var{ProtocolName}, \var{RecMode} and \var{TraceNames}.}
}}

\seealso{
\linkS4class{PRecording}
}
