% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inspect.R, R/Plot.R, R/PlotQC.R,
%   R/PlotRecording.R
\name{Inspect}
\alias{Inspect}
\alias{Plot}
\alias{PlotStimResp}
\alias{PlotTimeSeries}
\alias{PlotGroupComparison}
\alias{PlotQC}
\alias{PlotRecording}
\title{Plotting methods}
\usage{
Inspect(object, what = "I.mon")

PlotStimResp(
  X,
  StimTrace = "V-mon",
  RespTrace = "I-mon",
  Time,
  fun = mean,
  ReturnPMObject = T
)

PlotTimeSeries(X, RespTrace = "I-mon", Time, fun = mean, ReturnPMObject = T)

PlotGroupComparison(
  X,
  Sweep,
  RespTrace = "I-mon",
  Time,
  fun = mean,
  ReturnPMObject = T
)

PlotQC(object)

PlotRecording(object)
}
\arguments{
\item{what}{Specifies a plot to be plotted. any ggplot inside the Plots slot can be named}

\item{X}{A \linkS4class{PRecording} or \linkS4class{PCollection} object}

\item{StimTrace}{The name of the Trace that contains the Stimulus}

\item{RespTrace}{The name of the Trace that contains the Response}

\item{Time}{either a range of time points to keep, or, if \code{TimeExclusive} is \code{TRUE}, then two particular time points}

\item{ReturnPMObject}{whether to return a PCollection or a Matrix.}

\item{Sweep}{Sweep to analyse for group comparison}
}
\value{
A \linkS4class{PCollection} with an item added to the Plots slot if \code{ReturnPMObject=T} or a \link[=ggplot2::ggplot]{ggplot}.
}
\description{
These methods create "typical" ephsy graphs like dose-response curves, time-series or point statistics.
}
\section{Functions}{
\itemize{
\item \code{Inspect}: This function plots an interactive \link[=plotly]{plotly} graph from a plot generated using a Plot* method.

\item \code{PlotStimResp}: This method builds a dose-response curve

\item \code{PlotTimeSeries}: This method builds a time Series plot

\item \code{PlotGroupComparison}: This method builds a boxplot for comparison between groups as stored in the \linkS4class{PCollection}.

\item \code{PlotQC}: This function creates a basic visualization of quality-relevant recording parameters in the PCollection

\item \code{PlotRecording}: This function creates a basic visualization for each Trace in a single \linkS4class{PRecording} object.
}}

