% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1Plot.R, R/Inspect.R, R/PlotQC.R,
%   R/PlotRecording.R
\name{Plot}
\alias{Plot}
\alias{PlotStimResp}
\alias{PlotStimResp,PRecording-method}
\alias{PlotStimResp,PCollection-method}
\alias{PlotTimeSeries}
\alias{PlotTimeSeries,PRecording-method}
\alias{PlotTimeSeries,PCollection-method}
\alias{PlotGroupComparison}
\alias{PlotGroupComparison,PCollection-method}
\alias{Inspect}
\alias{Inspect,PRecording-method}
\alias{Inspect,PCollection-method}
\alias{PlotQC}
\alias{PlotQC,PCollection-method}
\alias{PlotRecording}
\alias{PlotRecording,PRecording-method}
\title{Plotting methods}
\usage{
PlotStimResp(
  X,
  StimTrace = "V-mon",
  RespTrace = "I-mon",
  Time,
  fun = mean,
  ReturnPMObject = T
)

\S4method{PlotStimResp}{PRecording}(
  X,
  StimTrace = "V-mon",
  RespTrace = "I-mon",
  Time,
  fun = mean,
  ReturnPMObject = T
)

\S4method{PlotStimResp}{PCollection}(
  X,
  StimTrace = "V-mon",
  RespTrace = "I-mon",
  Time,
  fun = mean,
  ReturnPMObject = T
)

PlotTimeSeries(X, RespTrace = "I-mon", Time, fun = mean, ReturnPMObject = T)

\S4method{PlotTimeSeries}{PRecording}(X, RespTrace = "I-mon", Time, fun = mean, ReturnPMObject = T)

\S4method{PlotTimeSeries}{PCollection}(X, RespTrace = "I-mon", Time, fun = mean, ReturnPMObject = T)

PlotGroupComparison(
  X,
  Sweep,
  RespTrace = "I-mon",
  Time,
  fun = mean,
  ReturnPMObject = T
)

\S4method{PlotGroupComparison}{PCollection}(
  X,
  Sweep,
  RespTrace = "I-mon",
  Time,
  fun = mean,
  ReturnPMObject = T
)

Inspect(X, what = "I.mon")

\S4method{Inspect}{PRecording}(X, what = "I.mon")

\S4method{Inspect}{PCollection}(X, what = "I.mon")

PlotQC(X)

\S4method{PlotQC}{PCollection}(X)

PlotRecording(X)

\S4method{PlotRecording}{PRecording}(X)
}
\arguments{
\item{X}{A \linkS4class{PRecording} or \linkS4class{PCollection} object}

\item{StimTrace}{The name of the \code{Trace} that contains the Stimulus}

\item{RespTrace}{The name of the \code{Trace} that contains the Response}

\item{Time}{either a range of time points to keep, or, if \code{TimeExclusive} is \code{TRUE}, then two particular time points}

\item{fun}{Function to apply on graph for stimulus response plotting}

\item{ReturnPMObject}{whether to return a \linkS4class{PRecording}/\linkS4class{PCollection} or a matrix.}

\item{Sweep}{Sweep to analyse for group comparison}

\item{what}{Specifies a plot to be plotted. any ggplot inside the Plots slot can be named}
}
\value{
A \linkS4class{PCollection} with an item added to the Plots slot if \code{ReturnPMObject=T} or a \link[ggplot2:ggplot]{ggplot}.
}
\description{
These methods create "typical" ephsy graphs like dose-response curves, time-series or point statistics.
}
\section{Functions}{
\itemize{
\item \code{PlotStimResp}: This method builds a dose-response curve

\item \code{PlotStimResp,PRecording-method}: Method for PRecording

\item \code{PlotStimResp,PCollection-method}: Method for PCollection

\item \code{PlotTimeSeries}: This method builds a time Series plot

\item \code{PlotTimeSeries,PRecording-method}: Method for PRecording

\item \code{PlotTimeSeries,PCollection-method}: Method for PCollection

\item \code{PlotGroupComparison}: This method builds a boxplot for comparison between groups as stored in the \linkS4class{PCollection}.

\item \code{PlotGroupComparison,PCollection-method}: Method for PCollection

\item \code{Inspect}: This function plots an interactive \link[=plotly]{plotly} graph from a plot generated using a Plot* method.

\item \code{Inspect,PRecording-method}: Method for PRecording

\item \code{Inspect,PCollection-method}: Method for PCollection

\item \code{PlotQC}: This function creates a basic visualization of quality-relevant recording parameters in the PCollection

\item \code{PlotQC,PCollection-method}: Method for PCollection

\item \code{PlotRecording}: This function creates a basic visualization for each Trace in a single \linkS4class{PRecording} object.

\item \code{PlotRecording,PRecording-method}: Method for PRecording
}}

