% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1Plot.R, R/BuildQCPlot.R,
%   R/BuildRecordingPlot.R, R/Inspect.R
\name{Plot}
\alias{Plot}
\alias{BuildStimRespPlot}
\alias{BuildTimeSeriesPlot}
\alias{BuildGroupComparisonPlot}
\alias{BuildQCPlot}
\alias{BuildRecordingPlot}
\alias{Inspect}
\title{(OK) Plotting methods}
\usage{
BuildStimRespPlot(
  X,
  StimTrace = "V-mon",
  RespTrace = "I-mon",
  Time,
  fun = mean,
  ReturnPMobject = T
)

BuildTimeSeriesPlot(
  X,
  RespTrace = "I-mon",
  Time,
  fun = mean,
  ReturnPMobject = T
)

BuildGroupComparisonPlot(
  X,
  Sweep,
  RespTrace = "I-mon",
  Time,
  fun = mean,
  ReturnPMobject = T
)

BuildQCPlot(X, ReturnPMobject = T)

BuildRecordingPlot(X)

Inspect(X, what = "I.mon")
}
\arguments{
\item{X}{A \linkS4class{PRecording} or \linkS4class{PCollection} object}

\item{StimTrace}{The name of the \code{trace} that contains the stimulus}

\item{RespTrace}{The name of the \code{trace} that contains the response}

\item{Time}{either a range of time points to keep, or, if \code{TimeExclusive} is \code{TRUE}, then two particular time points}

\item{fun}{Function to apply on graph for stimulus response plotting}

\item{ReturnPMobject}{whether to return a \linkS4class{PRecording}/\linkS4class{PCollection} or a matrix.}

\item{Sweep}{Sweep to analyse for group comparison}

\item{what}{Specifies a plot to be plotted. any ggplot inside the Plots slot can be named}
}
\value{
A \linkS4class{PCollection} with an item added to the Plots slot if \code{ReturnPMobject=T} or a \link[ggplot2:ggplot]{ggplot}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr
These methods create "typical" ephsy graphs like dose-response curves, time-series or point statistics.
}
\section{Functions}{
\itemize{
\item \code{BuildStimRespPlot}: This method builds a dose-response curve

\item \code{BuildTimeSeriesPlot}: This method builds a time Series plot

\item \code{BuildGroupComparisonPlot}: This method builds a boxplot for comparison between groups as stored in the \linkS4class{PCollection}.

\item \code{BuildQCPlot}: This function creates a basic visualization of quality-relevant recording parameters in the PCollection

\item \code{BuildRecordingPlot}: This function creates a basic visualization for each Trace in a single \linkS4class{PRecording} object.

\item \code{Inspect}: This function plots an interactive \link[=plotly]{plotly} graph from a plot generated using a Build* method.
}}

\examples{
\dontrun{
# return ggplot
BuildTimeSeriesPlot(aPRecording, RespTrace = "I-mon", Time = c(0.8,1), fun = mean, ReturnPMobject = F)
# return PRecording
aPRecording<-BuildTimeSeriesPlot(aPRecording, RespTrace = "I-mon", Time = c(0.8,1), fun = mean, ReturnPMobject = T)
Inspect(aPRecording,"TimeSeriesPlot")
}
}
