% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildPlot.R
\name{Plots}
\alias{Plots}
\alias{DoseRespPlot}
\alias{TimeSeriesPlot}
\alias{GroupComparisonPlot}
\title{Plotting methods}
\usage{
DoseRespPlot(
  X,
  StimTrace = "V-mon",
  RespTrace = "I-mon",
  Time,
  fun = mean,
  ReturnPMObject = T
)

TimeSeriesPlot(X, RespTrace = "I-mon", Time, fun = mean, ReturnPMObject = T)

GroupComparisonPlot(
  X,
  Sweep,
  RespTrace = "I-mon",
  Time,
  fun = mean,
  ReturnPMObject = T
)
}
\arguments{
\item{X}{a \link[=PMRecording]{PMRecording} or \link[=PMCollection]{PMCollection} Object}

\item{StimTrace}{The name of the Trace that contains the Stimulus}

\item{RespTrace}{The name of the Trace that contains the Response}

\item{Time}{either a range of time points to keep, or, if \code{TimeExclusive} is \code{TRUE}, then two particular time points}

\item{ReturnPMObject}{whether to return a PMCollection or a Matrix.}

\item{Sweep}{Sweep to analyse for group comparison}
}
\value{
A \linkS4class{PMCollection} with an item added to the Plots slot if \code{ReturnPMObject=T} or a \link[=ggplot2::ggplot]{ggplot}.
}
\description{
These methods create "typical" ephsy graphs like dose-response curves, time-series or point statistics.
}
\section{Functions}{
\itemize{
\item \code{DoseRespPlot}: This method builds a dose-response curve

\item \code{TimeSeriesPlot}: This method builds a time Series plot

\item \code{GroupComparisonPlot}: This method builds a boxplot for comparison between groups as stored in the \linkS4class{PMCollection}.
}}

