% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildPlot.R, R/Inspect.R
\name{Plots}
\alias{Plots}
\alias{PlotDoseResp}
\alias{PlotTimeSeries}
\alias{PlotGroupComparison}
\alias{Plot_Inspect}
\title{Plotting methods}
\usage{
PlotDoseResp(
  X,
  StimTrace = "V-mon",
  RespTrace = "I-mon",
  Time,
  fun = mean,
  ReturnPMObject = T
)

PlotTimeSeries(X, RespTrace = "I-mon", Time, fun = mean, ReturnPMObject = T)

PlotGroupComparison(
  X,
  Sweep,
  RespTrace = "I-mon",
  Time,
  fun = mean,
  ReturnPMObject = T
)

Plot_Inspect(object, what = "I.mon")
}
\arguments{
\item{X}{a \link[=PMRecording]{PMRecording} or \link[=PMCollection]{PMCollection} Object}

\item{StimTrace}{The name of the Trace that contains the Stimulus}

\item{RespTrace}{The name of the Trace that contains the Response}

\item{Time}{either a range of time points to keep, or, if \code{TimeExclusive} is \code{TRUE}, then two particular time points}

\item{ReturnPMObject}{whether to return a PMCollection or a Matrix.}

\item{Sweep}{Sweep to analyse for group comparison}

\item{what}{Specifies a plot to be plotted. any ggplot inside the Plots slot can be named}
}
\value{
A \linkS4class{PMCollection} with an item added to the Plots slot if \code{ReturnPMObject=T} or a \link[=ggplot2::ggplot]{ggplot}.
}
\description{
These methods create "typical" ephsy graphs like dose-response curves, time-series or point statistics.
}
\section{Functions}{
\itemize{
\item \code{PlotDoseResp}: This method builds a dose-response curve

\item \code{PlotTimeSeries}: This method builds a time Series plot

\item \code{PlotGroupComparison}: This method builds a boxplot for comparison between groups as stored in the \linkS4class{PMCollection}.

\item \code{Plot_Inspect}: This function plots an interactive \link[=plotly]{plotly} graph from a plot generated using a Build* method.
}}

