% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMRecordingParams.R
\docType{class}
\name{PMRecordingParams-class}
\alias{PMRecordingParams-class}
\alias{PMRecordingParams}
\title{S4 class storing Recording parameters for impoted PatchMaster Traces.}
\description{
This class stores imported PatchMaster Traces. Currently only checked for time series. It is strictly validated to assure data consistency
}
