% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMTrace.R
\docType{class}
\name{PMSeries-class}
\alias{PMSeries-class}
\alias{PMSeries}
\title{S4 class storing imported PatchMaster Traces.}
\description{
This class stores imported PatchMaster Traces. Currently only checked for time series. It is strictly validated to assure data consistency
}
\details{
\describe{
   \item{Traces}{Character vector containing names of the Traces (=Monitors) imported form the dat file and any subsequently computed Trace. Computed Traces have the same dimensions as imported}

\item{Units}{Character vector containing the SI units for data stored in corresponding Trace. Order as in Traces.}

\item{TimeTrace}{Numeric vector containing the time points of the recording.}

\item{Sweeps}{Ordered vector containing the names of the sweeps.}

\item{SweepTimes}{Numeric vector containing start times corresponding to Sweeps}

\item{Data}{List of matrices. One list item per Trace. Matrix rows correspond to TimeTrace, columns to Sweeps}

\item{Plots}{List that can contain any ggplot derived from the data. List item names that equal Traces are reserved.}

\item{metaData}{List that can contain any ggplot derived from the data. List item names that equal Traces are reserved.}

\item{RecordingParams}{An item of class PMRecordingParams containing recording parameters for that trace.}
 }
}
