% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodage.R
\name{recoder_champs}
\alias{recoder_champs}
\title{Recoder les champs d'une table a partir d'une expression VRAI/FAUX}
\usage{
recoder_champs(table, table_recodage, filtre = NULL, champs_table = TRUE)
}
\arguments{
\item{table}{La table à recoder.}

\item{table_recodage}{La table de recodage.}

\item{filtre}{La valeur de filtre.}

\item{champs_table}{Recodage réalisé uniquement sur le schamps déjà déjà présents dans la table.}
}
\value{
La table recodée
}
\description{
Recoder les champs d'une table à partir d'une expression VRAI/FAUX.
}
