% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recoder_champs}
\alias{recoder_champs}
\title{Recoder les champs d'une table a partir d'une expression VRAI/FAUX}
\usage{
recoder_champs(table, table_recodage, source = NULL, filtre = NULL,
  creer_champs = TRUE)
}
\arguments{
\item{table}{La table à recoder.}

\item{table_recodage}{La table de recodage.}

\item{source}{Nom de la source à filtrer dans la table \code{table_recodage}.}

\item{filtre}{La valeur de filtre.}

\item{creer_champs}{Créer les champs présents dans la table de recodage ou de niveaux mais absents dans la table à recoder.}
}
\value{
La table recodée
}
\description{
Recoder les champs d'une table à partir d'une expression VRAI/FAUX.
}
\examples{
# Recodage sans expression
patchr::recoder_champs(
table = dplyr::tibble(test = 1L),
table_recodage = dplyr::tibble(expression = NA_character_, champ = "test", valeur = "2L")
)

# Recodage avec expression
patchr::recoder_champs(
table = dplyr::tibble(test = c(1L, 2L)),
table_recodage = dplyr::tibble(expression = "test == 2", champ = "test", valeur = "3L")
)

}
