% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodage.R
\name{recoder_champs}
\alias{recoder_champs}
\title{Recoder les champs d'une table a partir d'une expression VRAI/FAUX}
\usage{
recoder_champs(table, table_recodage, filtre = NULL)
}
\arguments{
\item{table}{La table à recoder.}

\item{table_recodage}{La table de recodage.}

\item{.champ_id}{Le nom du champ "identifiant".}
}
\value{
La table recodée
}
\description{
Recoder les champs d'une table à partir d'une expression VRAI/FAUX.
}
