% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoder.R
\name{recoder_factor}
\alias{recoder_factor}
\title{recoder_factor}
\usage{
recoder_factor(table, table_recodage, source = NULL, filtre = NULL)
}
\arguments{
\item{table}{La table à recoder.}

\item{table_recodage}{La table de recodage.}

\item{source}{Nom de la source à filtrer dans la table \code{table_recodage}.}

\item{filtre}{La valeur de filtre.}
}
\value{
La table recodée
}
\description{
recoder_factor
}
