% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoder.R
\name{recoder_individu}
\alias{recoder_individu}
\title{Recoder les champs d'une table a partir d'un identifiant d'individu}
\usage{
recoder_individu(table, table_recodage, source = NULL,
  .champ_id = "code_etudiant")
}
\arguments{
\item{table}{La table à recoder.}

\item{table_recodage}{La table de recodage.}

\item{source}{Nom de la source à filtrer dans la table \code{table_recodage}.}

\item{.champ_id}{Le nom du champ contenant l'identifiant de l'individu.}
}
\value{
La table recodée
}
\description{
Recoder les champs d'une table à partir d'un identifiant d'individu.
}
