% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{remove_duplicate}
\alias{remove_duplicate}
\title{Remove duplicate rows from a table according a colname.}
\usage{
remove_duplicate(data, colname)
}
\arguments{
\item{data}{A data frame.}

\item{colname}{A column name for which duplicate rows are removed.}
}
\value{
A data frame with no duplicate for colname.
}
\description{
Remove duplicate rows from a table according a colname.
}
