% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetData.R
\name{SubsetData,PMSeries-method}
\alias{SubsetData,PMSeries-method}
\title{Subset a PMSeries object}
\usage{
\S4method{SubsetData}{PMSeries}(
  object,
  Traces = getTraces(object),
  Sweeps = getSweeps(object),
  Time = range(getTimeTrace(object)),
  TimeExclusive = F,
  nowarnings = F
)
}
\arguments{
\item{object}{a \link[=PMSeries]{PMSeries} or \link[=PMExperiment]{PMExperiment} object}

\item{Traces, Sweeps}{List of Traces/Sweeps to keep}

\item{Time}{either a range of time points to keep, or, if \code{TimeExclusive} is \code{TRUE}, then two particular time points}

\item{TimeExclusive}{Keep only the two time points stated under Time, not the range}

\item{Series}{Subset by Series name. Only for \link[=PMExperiment]{PMExperiment}.}

\item{Group}{Subset by Group name. Only for \link[=PMExperiment]{PMExperiment}.}
}
\value{
A \link[=PMSeries]{PMSeries} object
}
\description{
This function subsets \link[=PMSeries]{PMSeries} or \link[=PMExperiment]{PMExperiment} objects by Trace, Sweep or Time
}
