% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toLong.R
\docType{methods}
\name{toLong}
\alias{toLong}
\title{toLong}
\usage{
\S4method{toLong}{PMSeries}(object)
}
\arguments{
\item{object}{a PMSeries object}
}
\description{
converts data from a PMSeries object into its long representation, analog to tidyR's gather
}
