% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcode.R
\name{transcode}
\alias{transcode}
\title{Change column classes in a data frame.}
\usage{
transcode(data, data_transcode)
}
\arguments{
\item{data}{A data frame.}

\item{data_transcode}{A correspondance between colmanes and new R classes.}
}
\value{
A transcoded data frame.
}
\description{
The data frame columns are transcoded according to a correspondance table.\cr
The correspondance table data_transcode must at least contains tow fields : \code{column} for the column name and \code{class} for the new R class.
}
\examples{
dplyr::tibble(var1 = "a", var2 = 1) \%>\%
  patchr::transcode(dplyr::tibble(column = "var2", class = "character"))

}
